### 2020-06-18 #2364 @whedon
[PRE REVIEW]: pyveg: A Python package for analysing the time evolution of patterned vegetation using Google Earth Engine
**Submitting author:** @samvanstroud (<a href="http://orcid.org/0000-0002-7969-0301">Samuel Van Stroud</a>)
**Repository:** <a href="https://github.com/alan-turing-institute/monitoring-ecosystem-resilience" target ="_blank">https://github.com/alan-turing-institute/monitoring-ecosystem-resilience</a>
**Version:** v1.0.0
