### 2020-06-24 307e98d
add func to get date strings for sub-ranges
