### 2020-06-24 92af301
add null for missing timepoints in vegetation time series
