### 2020-06-24 b1ce935
remove input_veg_sequence and input_weather_seqeunce from list of required params
