### 2020-06-24 944e08d
move creation of container, or output directory, to BaseModule.run
