+++ pyveg/src/download_modules.py
@@ -13,14 +13,15 @@ from geetools import cloud_mask
 import cv2 as cv
 
 import ee
+
 ee.Initialize()
 
 from pyveg.src.date_utils import (
     find_mid_period,
     get_num_n_day_slices,
     slice_time_period_into_n,
-    slice_time_period
-    )
+    slice_time_period,
+)
 from pyveg.src.file_utils import download_and_unzip
 from pyveg.src.coordinate_utils import get_region_string
 from pyveg.src.gee_interface import apply_mask_cloud, add_NDVI
@@ -29,7 +30,8 @@ from pyveg.src.pyveg_pipeline import BaseModule
 
 # silence google API WARNING
 import logging
-logging.getLogger('googleapiclient.discovery_cache').setLevel(logging.ERROR)
+
+logging.getLogger("googleapiclient.discovery_cache").setLevel(logging.ERROR)
 
 
 class DownloaderModule(BaseModule):
@@ -38,19 +40,21 @@ class DownloaderModule(BaseModule):
     types of data, so we put it in this base class, and have data-type-specific
     code in subclasses.
     """
+
     def __init__(self, name):
         super().__init__(name)
         # some parameters that are common to all Downloaders
-        self.params +=  [("collection_name", [str]),
-                         ("coords", [list, tuple]),
-                         ("date_range", [list, tuple]),
-                         ("region_size", [float]),
-                         ("scale", [int]),
-                         ("output_location", [str]),
-                         ("output_location_type", [str]),
-                         ("replace_existing_files", [bool]) ]
+        self.params += [
+            ("collection_name", [str]),
+            ("coords", [list, tuple]),
+            ("date_range", [list, tuple]),
+            ("region_size", [float]),
+            ("scale", [int]),
+            ("output_location", [str]),
+            ("output_location_type", [str]),
+            ("replace_existing_files", [bool]),
+        ]
         return
-
 
     def set_default_parameters(self):
         """
@@ -72,7 +76,6 @@ class DownloaderModule(BaseModule):
 
         return
 
-
     def set_output_location(self, output_location=None):
         """
         If provided an output directory name, set it here,
@@ -87,14 +90,20 @@ class DownloaderModule(BaseModule):
             self.output_location = output_location[0]
 
         elif ("coords" in vars(self)) and ("collection_name" in vars(self)):
-            self.output_location = f'gee_{self.coords[0]}_{self.coords[1]}'\
-                +"_"+self.collection_name.replace('/', '-')
+            self.output_location = (
+                f"gee_{self.coords[0]}_{self.coords[1]}"
+                + "_"
+                + self.collection_name.replace("/", "-")
+            )
 
         else:
-            raise RuntimeError("""
+            raise RuntimeError(
+                """
             {}: need to set collection_name and coords before calling set_output_location()
-            """.format(self.name))
-
+            """.format(
+                    self.name
+                )
+            )
 
     def prep_data(self, date_range):
         """
@@ -112,37 +121,34 @@ class DownloaderModule(BaseModule):
         -------
         url_list:  a list of URLs from which zipfiles can be downloaded from GEE.
         """
-        region  = get_region_string(self.coords, self.region_size)
+        region = get_region_string(self.coords, self.region_size)
         start_date, end_date = date_range
 
         image_coll = ee.ImageCollection(self.collection_name)
         geom = ee.Geometry.Point(self.coords)
 
-        dataset = image_coll.filterBounds(geom).filterDate(start_date,end_date)
+        dataset = image_coll.filterBounds(geom).filterDate(start_date, end_date)
         dataset_size = dataset.size().getInfo()
 
         if dataset_size == 0:
-            print('No images found in this date rage, skipping.')
[…]
