+++ pyveg/scripts/download_gee_data.py
@@ -57,14 +57,17 @@ def main():
     """
 
     # crate argparse
-    parser = argparse.ArgumentParser(description="download from EE", epilog=help_text, 
-                                     formatter_class=argparse.RawDescriptionHelpFormatter)
+    parser = argparse.ArgumentParser(
+        description="download from EE",
+        epilog=help_text,
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+    )
 
-    # add arguments - keep it minimal here, but overwrite value from config 
+    # add arguments - keep it minimal here, but overwrite value from config
     # if option is given here
-    parser.add_argument("--start_date",help="YYYY-MM-DD")
-    parser.add_argument("--end_date",help="YYYY-MM-DD")
-    parser.add_argument("--coordinates",help="'long,lat'")
+    parser.add_argument("--start_date", help="YYYY-MM-DD")
+    parser.add_argument("--end_date", help="YYYY-MM-DD")
+    parser.add_argument("--coordinates", help="'long,lat'")
 
     args = parser.parse_args()
 
@@ -74,50 +77,56 @@ def main():
 
     # overwrite coords if specified
     if args.coordinates is not None:
-        coordinates = args.coordinates.split(',')
+        coordinates = args.coordinates.split(",")
         config.coordinates = (float(coordinates[0]), float(coordinates[1]))
 
-    # parse output directory 
-    config.output_dir += '__' + time.strftime("%Y-%m-%d_%H-%M-%S") 
-    config.output_dir = os.path.join('output', config.output_dir) # force into .gitignore
+    # parse output directory
+    config.output_dir += "__" + time.strftime("%Y-%m-%d_%H-%M-%S")
+    config.output_dir = os.path.join(
+        "output", config.output_dir
+    )  # force into .gitignore
 
     # before we run anythin, save the current config to the output dir
     if not os.path.exists(config.output_dir):
         os.makedirs(config.output_dir, exist_ok=True)
-    copyfile(config.__file__, os.path.join(config.output_dir, 'config_cached.py'))
+    copyfile(config.__file__, os.path.join(config.output_dir, "config_cached.py"))
 
     # print which collections we are running with
-    print('-'*35)
-    print('Running download_gee_data.py')
-    print('-'*35)
+    print("-" * 35)
+    print("Running download_gee_data.py")
+    print("-" * 35)
 
     # print job configuration to the user
-    print('Job Configuration:')
+    print("Job Configuration:")
 
     # search coordinate store to see if we have an entry for these coordinates
     long, lat = config.coordinates
-    entry = coordinate_store[(coordinate_store['longitude'] == long) & (coordinate_store['latitude'] == lat)]
+    entry = coordinate_store[
+        (coordinate_store["longitude"] == long) & (coordinate_store["latitude"] == lat)
+    ]
 
     if entry.empty:
-        print(f'* Warning: running with unknown coordinates: {config.coordinates}.')
+        print(f"* Warning: running with unknown coordinates: {config.coordinates}.")
     elif len(entry) == 1:
-        print(f'* Running with location id {entry.index[0]}:')
+        print(f"* Running with location id {entry.index[0]}:")
         print(entry.to_string(index=False))
     else:
-        raise RuntimeError('Coordinate store error.')
-    
-    print(f'* date range:\t {config.date_range}')
-    collection_names = [c['collection_name'] for c in config.data_collections.values()]
-    print(f'* collections:\t {collection_names}\n')
+        raise RuntimeError("Coordinate store error.")
+
+    print(f"* date range:\t {config.date_range}")
+    collection_names = [c["collection_name"] for c in config.data_collections.values()]
+    print(f"* collections:\t {collection_names}\n")
 
     # run!
[…]
