+++ pyveg/coordinates.py
@@ -19,72 +19,278 @@ In future think about:
 import pandas as pd
 
 # initialise a DataFrame to store coordinate
-coordinate_store = pd.DataFrame(columns=['continent', 'country', 'type', 'latitude', 'longitude'])
+coordinate_store = pd.DataFrame(
+    columns=["continent", "country", "type", "latitude", "longitude"]
+)
 
 # rename index
-coordinate_store.index.names = ['id']
+coordinate_store.index.names = ["id"]
 
 # --------------------------------------------------------------------------------
 # sudan
-coordinate_store.loc['00'] = {'continent': 'Africa', 'country': 'Sudan', 'type': 'spots',      'latitude': 11.58, 'longitude': 27.94}
-coordinate_store.loc['01'] = {'continent': 'Africa', 'country': 'Sudan', 'type': 'labyrinths', 'latitude': 11.12, 'longitude': 28.37}
-coordinate_store.loc['02'] = {'continent': 'Africa', 'country': 'Sudan', 'type': 'gaps',       'latitude': 10.96, 'longitude': 28.20}
+coordinate_store.loc["00"] = {
+    "continent": "Africa",
+    "country": "Sudan",
+    "type": "spots",
+    "latitude": 11.58,
+    "longitude": 27.94,
+}
+coordinate_store.loc["01"] = {
+    "continent": "Africa",
+    "country": "Sudan",
+    "type": "labyrinths",
+    "latitude": 11.12,
+    "longitude": 28.37,
+}
+coordinate_store.loc["02"] = {
+    "continent": "Africa",
+    "country": "Sudan",
+    "type": "gaps",
+    "latitude": 10.96,
+    "longitude": 28.20,
+}
 
 # niger
-coordinate_store.loc['03'] = {'continent': 'Africa', 'country': 'Niger', 'type': 'tiger bush', 'latitude': 13.12, 'longitude': 2.59}
-coordinate_store.loc['04'] = {'continent': 'Africa', 'country': 'Niger', 'type': 'tiger bush', 'latitude': 13.17, 'longitude': 1.58}
+coordinate_store.loc["03"] = {
+    "continent": "Africa",
+    "country": "Niger",
+    "type": "tiger bush",
+    "latitude": 13.12,
+    "longitude": 2.59,
+}
+coordinate_store.loc["04"] = {
+    "continent": "Africa",
+    "country": "Niger",
+    "type": "tiger bush",
+    "latitude": 13.17,
+    "longitude": 1.58,
+}
 
 # senegal
-coordinate_store.loc['05'] = {'continent': 'Africa', 'country': 'Senegal', 'type': 'labyrinths', 'latitude': 15.20, 'longitude': -15.20}
-coordinate_store.loc['06'] = {'continent': 'Africa', 'country': 'Senegal', 'type': 'labyrinths', 'latitude': 15.09, 'longitude': -15.04}
-coordinate_store.loc['07'] = {'continent': 'Africa', 'country': 'Senegal', 'type': 'gaps',       'latitude': 15.80, 'longitude': -14.36}
-coordinate_store.loc['08'] = {'continent': 'Africa', 'country': 'Senegal', 'type': 'gaps',       'latitude': 15.11, 'longitude': -14.53}
+coordinate_store.loc["05"] = {
+    "continent": "Africa",
+    "country": "Senegal",
+    "type": "labyrinths",
+    "latitude": 15.20,
+    "longitude": -15.20,
+}
+coordinate_store.loc["06"] = {
+    "continent": "Africa",
+    "country": "Senegal",
+    "type": "labyrinths",
+    "latitude": 15.09,
+    "longitude": -15.04,
+}
+coordinate_store.loc["07"] = {
[…]
