### 2020-06-24 c314257
remove old download scripts
