### 2020-06-29 a5b1cf5
better reporting of success/failure for all Modules
