+++ pyveg/src/batch_utils.py
@@ -319,6 +319,15 @@ def create_job(job_id, pool_id=None, batch_service_client=None):
     batch_service_client.job.add(job)
 
 
+def delete_job(job_id, batch_service_client=None):
+    """
+    Removes a job, and associated tasks.
+    """
+    if not batch_service_client:
+        batch_service_client = create_batch_client()
+    batch_service_client.job.delete(job_id)
+
+
 def create_batch_client():
     credentials = batch_auth.SharedKeyCredentials(
         config["batch_account_name"], config["batch_account_key"]
