### 2020-07-01 #2364 @samvanstroud
Hi @danielskatz, thanks for writing. A license exists, though it is in the pyveg/ directory, as really the submission is only for the code in this dir (https://github.com/alan-turing-institute/monitoring-ecosystem-resilience/blob/master/pyveg/LICENSE.md). (We are developing in parallel R code which is in the rveg/ dir, however this is not in a state of readiness to be submitted to JOSS at this time).

Perhaps this is why the bot fails to detect a license? Please advise if I should move the license to the top level of the repo. Another option if needed is to make a new repo, and strip out everything that is not strictly necessary for the core pyveg package we are submitting for review. Just let me know if this option would make more sense. Thank you!
