### 2020-07-01 #2364 @danielskatz
> Perhaps this is why the bot fails to detect a license? Please advise if I should move the license to the top level of the repo.

Yes.  Please do move it unless there's some reason not to.

> Another option if needed is to make a new repo, and strip out everything that is not strictly necessary for the core pyveg package we are submitting for review. Just let me know if this option would make more sense.

As long as the material to be reviewed is clear, a repo with additional content is ok, though note that at the end of the process, we will archive the repo along with publishing the paper, so I think this is up to you.
