### 2020-07-01 #2364 @samvanstroud
Thanks for the advice @danielskatz, the license has been moved. For now I think we'll otherwise leave the repo as is, and will also tag package authors @crangelsmith @nbarlowATI in case they want to follow this thread or add any info.
