### 2020-07-02 9f6c1d5
Check whether batch tasks have failed dependencies
