### 2020-07-02 e8fb5cb
function that creates markdown and pdf document based on analysis plots (#326)
