### 2020-07-02 a2b32c8
making ews plots wider for aesthetic reasons (#326)
