### 2020-07-02 d08e068
adding pandoc and mdutils to requirement files (#326)
