+++ pyveg/requirements.txt
@@ -20,3 +20,5 @@ imageio
 opencv-python
 statsmodels
 ewstools
+pypandoc
+mdutils
