### 2020-07-03 a2a6bc2
including reporting in the analysis script (#326)
