+++ pyveg/scripts/analyse_gee_data.py
@@ -38,6 +38,7 @@ from pyveg.src.plotting import (
     kendall_tau_histograms
 )
 
+from pyveg.scripts.create_analysis_report import create_markdown_pdf_report
 
 def run_time_series_analysis(filename, output_dir, detrended=False):
     """
@@ -257,6 +258,17 @@ def analyse_gee_data(input_dir, spatial):
                 data_df_geo = convert_to_geopandas(df.copy())
                 data_df_geo_coarse = coarse_dataframe(data_df_geo.copy(), 2)
                 create_lat_long_metric_figures(data_df_geo_coarse, 'offset50', spatial_subdir)
+    # ------------------------------------------------
+    # reporting
+    print('\nAnalysis complete.\n')
+
+    print('\nCreating report.\n')
+
+    for collection_name, df in dfs.items():
+        if collection_name == 'COPERNICUS/S2' or 'LANDSAT' in collection_name:
+
+            create_markdown_pdf_report(input_dir,collection_name)
+
     # ------------------------------------------------
 
     print('\nAnalysis complete.\n')
