### 2020-07-03 04e543c
adapting the reporting function to work easier with the gee analysis (#326)
