### 2020-07-03 f09c9a4
add a protection to avoid failure in the analysis side in case some figures are missing (#326)
