### 2020-07-03 2a94b58
adding a protection for the case that name of the directory doesn't have the usual nomenclaure of "gee_coords_satellite" (#326)
