### 2020-07-03 d9dfa50
adding pandoc to requirements.txt (#326)
