+++ pyveg/requirements.txt
@@ -22,3 +22,4 @@ statsmodels
 ewstools
 pypandoc
 mdutils
+pandoc
