### 2020-07-03 35b41cf
catching an error to fix CI and pandoc problem (#326)
