+++ pyveg/coordinates.py
@@ -308,7 +308,14 @@ coordinate_store.loc['35'] = {
   'longitude': 119.89
 }
 
-
+# Burkina Faso
+coordinate_store.loc["36"] = {
+    "continent": "Africa",
+    "country": "Burkina Faso",
+    "type": "tiger bush",
+    "latitude": -0.68,
+    "longitude": 14.92,
+}
 
 
 
