### 2020-07-04 #2364 @usethedata
@samvanstroud A couple of thoughts as I walk through this in preparation for finding reviewers.  I think these are things to address prior to review, but this is a discussion about subjects where there are multiple points of view, so please do provide your perspective where you see things differently.

1. There are multiple sources of NDVI in Google Earth Engine.  I believe the paper itself should be clearer about which NDVI is being used.  It appears you've preconfigured for Sentinal-2, and Landsat[4578].  My opinion (colored by the fact that I work on the NASA Earth Observing System) is that the _primary_ sources for that NDVI data should also be cited, in addition to citing GEE and the services it provides.  This is an interesting challenge in and around GEE, in that they make a copy of the data (with some lag relative to updates at the primary data source) and provide some very useful services on top of that data, but the primary source(s) for that data are rarely credited.

1. Likewise, I suggest citing the primary source for the ERA5 weather collection.

1. I'm curious why you chose to store the coordinate locations in a python file rather than in a separate data file (e.g. JSON), and then refer to the locations by a number, rather than a location name. This seems something likely to cause confusion and inadvertent error.

1. The significance/rationale for the specific 33 locations pre-configured in the package is unclear to me, based on reading the paper and a 20-30 minute examination of the source code.  A minor nit is that location 13, listed as being in the USA is actually in Mexico.  I suggest that a sentence or two about the preconfigured locations and any factors involved in extending this list is appropriate for the paper and/or the documentation.
