### 2020-07-04 #2364 @usethedata
Disclaimer: I'm the principal investigator for the additional information I'm commenting on.  I'm posting it here publicly to make it very clear that this is simply information provided colleague to colleague.  It is something that I think is relevant to the science domain in which the authors are working and their potential future work.  It is not directly relevant to this particular paper or software.

One of the very valuable things that GEE brings to the ecologist's toolbox is very simple subsetting.  For NDVI from the MODIS/VIIRS platforms, my group provides REST web services that provide subsets in easy-to-use formats.  See https://modis.ornl.gov/data/modis_webservice.html for additional information.  Those services also provide weather information (for North America, only, unfortunately) from the Daymet data product.
