+++ pyveg/paper/paper.bib
@@ -73,3 +73,15 @@
     year = {2020},
     url={https://github.com/ThomasMBury/ewstools}
 }
+
+@misc{era5,
+    title={Copernicus Climate Change Service (C3S): ERA5: Fifth generation of ECMWF atmospheric reanalyses of the global climate.}
+}
+
+@misc{sentinel2,
+    title={Contains modified Copernicus Sentinel data, courtesy of ESA}
+}
+
+@misc{landsat,
+    title={Landsat-4, Landsat-5, Landsat-7, Landsat-8 data courtesy of the U.S. Geological Survey}
+}
