+++ pyveg/paper/paper.bib
@@ -65,6 +65,7 @@
 
 @misc{geetools,
     year = {2020},
+    author={E. Principe, Rodrigo et al},
     url={https://github.com/gee-community/gee_tools}
 }
 
@@ -75,13 +76,19 @@
 }
 
 @misc{era5,
+    author={ECMWF},
+    year = {2020},
     title={Copernicus Climate Change Service (C3S): ERA5: Fifth generation of ECMWF atmospheric reanalyses of the global climate.}
 }
 
 @misc{sentinel2,
+    author={ESA},
+    year = {2020},
     title={Contains modified Copernicus Sentinel data, courtesy of ESA}
 }
 
 @misc{landsat,
+    author={USGS},
+    year = {2020},
     title={Landsat-4, Landsat-5, Landsat-7, Landsat-8 data courtesy of the U.S. Geological Survey}
 }
