+++ pyveg/paper/paper.bib
@@ -65,7 +65,7 @@
 
 @misc{geetools,
     year = {2020},
-    author={E. Principe, Rodrigo et al},
+    author={Principe, R.E. et al},
     url={https://github.com/gee-community/gee_tools}
 }
 
