### 2020-07-06 6c090f8
merge paper from master
