### 2020-07-06 #2364 @samvanstroud
Hello @usethedata. Firstly thank you for editing this submission and for taking a first look. Here are some replies to your comments:

1 & 2. Thanks for pointing this out, we agree entirely the primary data source should be acknowledged. I've just updated the paper with acknowledgements for Landsat, Sentinel-2, and ERA5. Let me know if something is still missing.

3. This is a good point, and your intuition is correct in that there is not really a good reason why we store this information in a Python file. However, I suspect we will keep some sort of numerical ID, as this is easier than defining a naming convention (we have multiple locations per country, for example, so country is not unique enough. Of course we could use the lat, long coordinates themselves, but this then effectively becomes as intelligible as a numerical ID for me at least). Perhaps a good middle ground would be to follow the following convention: "CountryName_01"?

4. The coordinates file is currently justified in the paper with the sentence: "An evolving list of PVP locations, obtained through both literature and manual searches, is included in the package at pyveg/coordinates.py.". This is really just meant to be an exhaustive list of all the interesting patterned vegetation locations we have managed to find, the coordinates of which may be of use to anyone studying patterned vegetation ecosystems, and also to interface with the code. If that's not clear let me know, and I can add another sentence or so to the paper to clear things up. Good spot for location 13! I'll change that now.

Thanks again for your great comments.
