### 2020-07-06 cfc2bc6
Use shutil.copyfile to copy downloaded file to output location
