+++ pyveg/src/pyveg_pipeline.py
@@ -21,6 +21,8 @@ import time
 
 from pyveg.src.file_utils import save_json
 
+from shutil import copyfile
+
 try:
     from pyveg.src import azure_utils
     from pyveg.src import batch_utils
@@ -440,14 +442,15 @@ class BaseModule(object):
                     if file_endings:
                         for ending in file_endings:
                             if filename.endswith(ending):
-                                subprocess.run(
-                                    [
-                                        "cp",
-                                        "-r",
-                                        os.path.join(root, filename),
-                                        os.path.join(output_location, filename),
-                                    ]
-                                )
+                                copyfile(os.path.join(root,filename),os.path.join(output_location,filename))
+                                #subprocess.run(
+                                 #   [
+                                        #"-r"
+                                #        "cp",
+                                #        os.path.join(root, filename),
+                                #        os.path.join(output_location, filename),
+                                #    ]
+                                #)
                     else:
                         subprocess.run(
                             [
