### 2020-07-08 6a9f989
add logging to file
