+++ pyveg/src/pyveg_pipeline.py
@@ -18,6 +18,26 @@ import json
 import subprocess
 import time
 
+import logging
+from logging.handlers import RotatingFileHandler
+
+logger = logging.getLogger("pyveg_logger")
+formatter = logging.Formatter("%(asctime)s [%(levelname)s] %(message)s")
+logger.setLevel(logging.INFO)
+
+c_handler = logging.StreamHandler()
+c_handler.setFormatter(formatter)
+f_handler = RotatingFileHandler(
+    "pyveg_{}.log".format(time.strftime("%Y-%m-%d_%H-%M-%S")),
+    maxBytes=5 * 1024 * 1024, backupCount=10
+)
+f_handler.setFormatter(formatter)
+
+logger.addHandler(f_handler)
+logger.addHandler(c_handler)
+
+
+
 
 from pyveg.src.file_utils import save_json
 
@@ -170,7 +190,7 @@ class Sequence(object):
 
     def set_config(self, config_dict):
         for k, v in config_dict.items():
-            print("{}: setting {} to {}".format(self.name, k, v))
+            logger.info("{}: setting {} to {}".format(self.name, k, v))
             self.__setattr__(k, v)
 
     def configure(self):
@@ -203,8 +223,8 @@ class Sequence(object):
         on which we depend, and if so, wait for them to finish.
         """
         if len(self.depends_on) > 0:
-            print(
-                "{} will check if all Sequences I depend on have finished".format(
+            logger.info(
+                "{} Checking if all dependency Sequences have finished".format(
                     self.name
                 )
             )
@@ -213,13 +233,14 @@ class Sequence(object):
                 num_seq_finished = 0
                 for seq_name in self.depends_on:
                     seq = self.parent.get(seq_name)
-                    print("{}: checking status of {}".format(self.name, seq.name))
+                    logger.info("{}: checking status of {}".format(self.name, seq.name))
                     if seq.check_if_finished():
-                        print("{}   ... finished".format(seq.name))
+                        logger.info("     {}   ... finished".format(seq.name))
                         num_seq_finished += 1
                     dependencies_finished = num_seq_finished == len(self.depends_on)
-                    print(
-                        "{} / {} dependencies finished".format(
+                    logger.info(
+                        "{}: {} / {} dependencies finished".format(
+                            self.name,
                             num_seq_finished, len(self.depends_on)
                         )
                     )
@@ -256,11 +277,11 @@ class Sequence(object):
         For all modules in the sequence, print out how many jobs
         succeeded or failed.
         """
-        print("\nSequence {}".format(self.name))
-        print("-----------------\n")
+        logger.info("\nSequence {}".format(self.name))
+        logger.info("-----------------\n")
         for module in self.modules:
             module.print_run_status()
-        print("\n")
+        logger.info("\n")
 
     def get(self, mod_name):
         """
@@ -285,7 +306,7 @@ class Sequence(object):
         if has_batch_job:
             self.batch_job_id = self.name + "_" + time.strftime("%Y-%m-%d_%H-%M-%S")
             batch_utils.create_job(self.batch_job_id)
-            print(
+            logger.info(
                 "Sequence {}: Creating batch job {}".format(
                     self.name, self.batch_job_id
                 )
@@ -300,11 +321,11 @@ class Sequence(object):
         num_modules_finished = 0
 
         for module in self.modules:
-            print("{}: checking status of {}".format(self.name, module.name))
+            logger.info("{}: checking status of {}".format(self.name, module.name))
             if module.check_if_finished():
-                print("{}   ... finished".format(module.name))
+                logger.info("{}   ... finished".format(module.name))
                 num_modules_finished += 1
-        print(
+        logger.info(
             "{} / {} modules finished".format(num_modules_finished, len(self.modules))
         )
 
[…]
