+++ pyveg/src/processor_modules.py
@@ -43,7 +43,7 @@ from pyveg.src.subgraph_centrality import (
 )
 from pyveg.src import azure_utils
 from pyveg.src import batch_utils
-from pyveg.src.pyveg_pipeline import BaseModule
+from pyveg.src.pyveg_pipeline import BaseModule, logger
 
 
 class ProcessorModule(BaseModule):
@@ -194,7 +194,7 @@ class ProcessorModule(BaseModule):
         """
         loop over dates and call process_single_date on all of them.
         """
-        print("{}: Running local".format(self.name))
+        logger.info("{}: Running local".format(self.name))
         if "dates_to_process" in vars(self) and len(self.dates_to_process) > 0:
             date_strings = self.dates_to_process
         else:
@@ -203,7 +203,7 @@ class ProcessorModule(BaseModule):
             )
 
         for date_string in date_strings:
-            print(
+            logger.info(
                 "date string {} input exists {} output exists {}".format(
                     date_string,
                     self.check_input_data_exists(date_string),
@@ -235,41 +235,41 @@ class ProcessorModule(BaseModule):
         if len(self.depends_on) > 0:
             for dependency in self.depends_on:
                 if not (self.parent and self.parent.get(dependency)):
-                    print(
+                    logger.info(
                         "{} couldn't retrieve dependency {}".format(
                             self.name.dependency
                         )
                     )
                     continue
                 dependency_module = self.parent.get(dependency)
-                print(
+                logger.info(
                     "{}: has dependency on {}".format(self.name, dependency_module.name)
                 )
                 if (
                     not "run_mode" in vars(dependency_module)
                 ) or dependency_module.run_mode == "local":
-                    print(
+                    logger.info(
                         "{}: dependency module {} is in local run mode".format(
                             self.name, dependency_module.name
                         )
                     )
                     continue
-                print(
+                logger.info(
                     "has {} submitted all tasks? {}".format(
                         dependency_module.name, dependency_module.all_tasks_submitted
                     )
                 )
                 while not dependency_module.all_tasks_submitted:
-                    print(
+                    logger.info(
                         "{}: waiting for {} to submit all batch tasks".format(
                             self.name, dependency_module.name
                         )
                     )
-                    print(".", end="")
+                    logger.info(".", end="")
                     sys.stdout.flush()
                     time.sleep(1)
                 task_dependencies.update(dependency_module.batch_task_dict)
-        print("{} return task_dependencies {}".format(self.name, task_dependencies))
+        logger.info("{} return task_dependencies {}".format(self.name, task_dependencies))
         return task_dependencies
 
     def create_task_dict(self, task_id, date_list, dependencies=[]):
@@ -294,7 +294,7 @@ class ProcessorModule(BaseModule):
         to pass to the batch_utils.submit_tasks function.
         """
 
-        print("{} running in batch".format(self.name))
+        logger.info("{} running in batch".format(self.name))
         self.all_tasks_submitted = False
         self.start_time = datetime.datetime.now()
         task_dicts = []
@@ -305,17 +305,17 @@ class ProcessorModule(BaseModule):
             date_strings = sorted(
                 self.list_directory(self.input_location, self.input_location_type)
             )
-            print(
+            logger.info(
                 "number of date strings in input location {}".format(len(date_strings))
             )
             date_strings = [
                 ds for ds in date_strings if self.check_input_data_exists(ds)
             ]
-            print("number of date strings with input data {}".format(len(date_strings)))
[…]
