### 2020-07-08 ac8b8e9
cleanup batch resources at end of sequence
