### 2020-07-09 7e02607
changing order on report after Nick suggestions (#326)
