### 2020-07-11 edb0407
fix crash in combiner modules if missing weather json
