### 2020-07-11 fb2d5ba
fix Sequence cleanup if run_mode not defined
