### 2020-07-13 #2364 @usethedata
I'm working on getting reviewers.  Requests sent out via email.

On the issue with the location identifiers, my particular background makes me aware of the kinds of issues that come up when people have to input information (like typing in a location number) and single character errors result in valid, but incorrect results.  By this, I mean that (for example) 08 and 09 are both valid, but different, entries. I've generally found it to be a best practice that any time I need a user to type something in off a list or put something into a configuration file that I set things up so that single character errors result in an invalid entry.  I did a six sigma study of this several years ago, and the error rate from technicians typing in a 3 digit number that they were reading off of a container was astounding.

So, your mileage may vary, and you've already done a lot of work on this.  But, something you might wish to consider for the future.  I realize that's a challenge for this particular challenge -- it's not like you've got something like LTER site names to work from, and there's not an obvious scheme.  So, no action needed -- just something you might consider for future work.
