### 2020-07-14 #2364 @nbarlowATI
Hi @usethedata
thanks, yes, this is an excellent point, and I have fallen victim to this myself (changing the number in one place but not the corresponding output location...).
For our production run for the paper we're working on, we plan to mitigate this by scripting the workflow, from generation of the config files that define inputs and outputs, to the upload of the results to an open source repository (possibly Zenodo), such that this `coordinates.py` file (which may indeed become a csv or json as you suggest) is the definitive and unique link from ID to location.
We believe that this way we can ensure at least that our inputs and outputs of the code are consistent, and any risks from transcription errors are pushed further down the line into paper drafting, where there will be more sets of eyes to check.

Thanks again for looking at this and the very useful feedback!
