### 2020-07-14 7ec233a
more useful error message for missing config file
