### 2020-07-14 4b3c5fc
fix crash  in check_task_failed_dependencies if task has no depends_on
