### 2020-07-15 897c9bc
fix case where we try to do os.listdir on a temporary file
