### 2020-07-15 41495d7
fix missing parameter to check_task_failed_dependencies
