### 2020-07-16 c624c29
change some output to 'debug'
