### 2020-07-16 47a9015
add config files for running some locations on batch
