### 2020-07-17 #2483 @whedon
[REVIEW]: pyveg: A Python package for analysing the time evolution of patterned vegetation using Google Earth Engine
**Submitting author:** @samvanstroud (<a href="http://orcid.org/0000-0002-7969-0301">Samuel Van Stroud</a>)
**Repository:** <a href="https://github.com/alan-turing-institute/monitoring-ecosystem-resilience" target ="_blank">https://github.com/alan-turing-institute/monitoring-ecosystem-resilience</a>
**Version:** 1.1.0

@arbennett , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
