### 2020-07-20 ba58a24
add new arguments for azure location
