+++ pyveg/src/zenodo_utils.py
@@ -16,7 +16,19 @@ with "sandbox.zenodo.org" in the URLs, and put the token into a file named "zeno
 then call the functions in this module with the "test" argument set to True.
 """
 import os
+import shutil
+import json
 import requests
+import tempfile
+from zipfile import ZipFile, BadZipFile
+
+from pyveg.src.file_utils import get_filepath_after_directory
+from pyveg.src.coordinate_utils import find_coords_string
+try:
+    import pyveg.zenodo_config as config
+except:
+    print("File zenodo_config.py not found - will not be able to access the Zenodo API")
+
 
 def get_base_url_and_token(test=False):
     """
@@ -31,14 +43,25 @@ def get_base_url_and_token(test=False):
     base_url: str, the first part of the URL for the API
     api_token: str, the personal access token, read from a file.
     """
-    token_dir = os.path.join(os.path.dirname(__file__), "..", "configs")
-    token_filename = os.path.join(token_dir, "zenodo_test_api_token") if test \
-        else os.path.join(token_dir, "zenodo_api_token")
-    if not os.path.exists(token_filename):
-        raise FileNotFoundError("Unable to find file containing Zenodo API token: {}".format(token_filename))
-    token = open(token_filename).read().strip()
-    base_url = 'https://sandbox.zenodo.org/api/' if test else 'https://zenodo.org/api/'
+    if test:
+        base_url = config.test_api_credentials["base_url"]
+        token = config.test_api_credentials["api_token"]
+    else:
+        base_url = config.prod_api_credentials["base_url"]
+        token = config.prod_api_credentials["api_token"]
     return base_url, token
+
+
+def get_deposition_id(test=False):
+    """
+    If we have previously created a deposition, we hopefully stored its ID in
+    the zenodo_config.py file.
+    """
+    if test:
+        return config.test_api_credentials["deposition_id"]
+    else:
+        return config.prod_api_credentials["deposition_id"]
+
 
 
 def list_depositions(test=False):
@@ -60,7 +83,6 @@ def list_depositions(test=False):
         print("Error retrieving depositions", r.content)
         return False
     return r.json()
-
 
 
 def create_deposition(test=False):
@@ -88,15 +110,60 @@ def create_deposition(test=False):
     return r.json()
 
 
+def get_deposition_info(deposition_id, test=False):
+    """
+    Get the JSON object containing details of a deposition.
 
-def upload_file(filename, bucket_url, test=False):
+    Parameters
+    ==========
+    deposition_id: int, ID of the deposition.
+    test: bool, if True use the sandbox API, if False will use the real one.
+
+    Returns
+    =======
+    dep_info: dict, information about the deposition
+    """
+    headers = {"Content-Type": "application/json"}
+    base_url, api_token = get_base_url_and_token(test)
+    params = {'access_token': api_token}
+    r = requests.get('{}/deposit/depositions/{}'.format(base_url, deposition_id),
+                   params=params,
+                   json={},
+                   headers=headers)
+    if r.status_code != 200:
+        print("Error getting deposition", r.content)
+        return {}
+    return r.json()
+
+
+def get_bucket_url(deposition_id, test=False):
+    """
+    For a given deposition_id, find the URL needed to upload a file.
+
+    Parameters
+    ==========
+    deposition_id: int, ID of the deposition.
+    test: bool, if True use the sandbox API, if False will use the real one.
+
+    Returns
+    =======
+    bucket_url: str, the URL of the bucket for this deposition, or empty string if id not found
+    """
+    dep_info = get_deposition_info(deposition_id, test)
+    if not dep_info:
+        print("Deposition {} not found".format(deposition_id))
+        return ""
[…]
