### 2020-07-24 6b205e9
add script and entrypoint for zenodo upload
