+++ setup.py
@@ -23,5 +23,6 @@ setup(
         "pyveg_gee_analysis=pyveg.scripts.analyse_gee_data:main",
         "pyveg_run_pipeline=pyveg.scripts.run_pyveg_pipeline:main",
         "pyveg_run_module=pyveg.scripts.run_pyveg_module:main"
+        "pyveg_zenodo_upload=pyveg.scripts.upload_to_zenodo:main"
     ]},
 )
