### 2020-07-24 14f17a2
automatic zenodo upload when running analysis
