### 2020-07-24 b97b66d
change return type of upload_file to bool for simplicity
