### 2020-07-24 e443bb5
put zenodo upload zipfile in tempdir
