### 2020-07-27 ac13127
adapt analysse_gee_data to be able to read summary json either locally or from azure
