+++ pyveg/scripts/analyse_gee_data.py
@@ -41,6 +41,11 @@ from pyveg.src.plotting import (
 from pyveg.scripts.create_analysis_report import create_markdown_pdf_report
 from pyveg.scripts.upload_to_zenodo import upload_results
 
+try:
+    from pyveg.src import azure_utils
+except:
+    print("Unable to import azure_utils")
+
 
 def run_time_series_analysis(filename, output_dir, detrended=False):
     """
@@ -203,72 +208,92 @@ def run_early_warnings_resilience_analysis(filename, output_dir):
             )
 
 
-def analyse_gee_data(input_dir, spatial=False, upload_to_zenodo=False, upload_to_zenodo_test=False):
+def analyse_gee_data(input_location,
+                     input_location_type,
+                     output_dir,
+                     do_time_series=True,
+                     do_spatial=False,
+                     upload_to_zenodo=False,
+                     upload_to_zenodo_test=False):
     """
     Run analysis on dowloaded gee data
 
     Parameters
     ----------
-    input_dir : string
-        Path to directory with downloaded dada
-    do_spatial_plot: bool
-        Option to run spatial analysis and do plots
-    do_time_series_plot: bool
+    input_location : str
+        Location of results_summary.json output from pyveg_run_pipeline
+    input_location_type: str
+        Can be 'local' or 'azure'.
+    output_dir: str,
+        Location for outputs of the analysis
+    do_time_series: bool
         Option to run time-series analysis and do plots
+    do_spatial: bool
+        Option to run spatial analysis and do plots
+    upload_to_zenodo: bool
+        Upload results to the production Zenodo repo
+    upload_to_zenodo_test: bool
+        Upload results to the sandbox Zenodo repo
     """
+
+    # read the results_summary.json
+    input_json = read_results_summary(input_location, input_location_type)
 
     # preprocess input data
     ts_dirname, dfs = preprocess_data(
-        input_dir, n_smooth=4, resample=False, period="MS"
+        input_json, output_dir, n_smooth=4, resample=False, period="MS"
     )
 
     # get filenames of preprocessed data time series
     ts_filenames = [f for f in os.listdir(ts_dirname) if "time_series" in f]
 
     # put all analysis results in this dir
-    output_dir = os.path.join(input_dir, "analysis")
-    if not os.path.exists(output_dir):
-        os.makedirs(output_dir, exist_ok=True)
+    output_analysis_dir = os.path.join(output_dir, "analysis")
+    if not os.path.exists(output_analysis_dir):
+        os.makedirs(output_analysis_dir, exist_ok=True)
 
     print("\nRunning Analysis...")
     print("-" * len("Running Analysis..."))
 
     # plot the feature vectors
-    plot_feature_vector(output_dir)
+    plot_feature_vector(output_analysis_dir)
 
+    # time-series analysis and plotting
+    # -----------------------------------
     # for each time series
-    for filename in ts_filenames:
+    if do_time_series:
+        for filename in ts_filenames:
 
-        ts_file = os.path.join(ts_dirname, filename)
-        print(f'\n* Analysing "{ts_file}"...')
-        print("." * 50)
+            ts_file = os.path.join(ts_dirname, filename)
+            print(f'\n* Analysing "{ts_file}"...')
+            print("." * 50)
 
-        # run the standard or detrended analysis
-        if "detrended" in filename:
-            output_subdir = os.path.join(output_dir, "detrended")
-            run_time_series_analysis(ts_file, output_subdir, detrended=True)
+            # run the standard or detrended analysis
+            if "detrended" in filename:
+                output_subdir = os.path.join(output_analysis_dir, "detrended")
[…]
