### 2020-07-27 fd0fb29
fix tests to take account of new interface for analysis preprocessing
