+++ pyveg/tests/test_analysis_preprocessing.py
@@ -1,19 +1,22 @@
 """
 Test the functions in analysis_preprocessing.py
 """
+import os
+import json
 
-from pyveg.src.analysis_preprocessing import *
+from pyveg.src.analysis_preprocessing import read_json_to_dataframes
 
 
 def test_read_json_to_dataframes():
-    test_df_dict = read_json_to_dataframes(
-        os.path.join(
+
+    json_filename = os.path.join(
             os.path.dirname(__file__),
             "..",
             "testdata",
             "network_json_data/test-results-summary.json",
-        )
     )
+    summary_json = json.load(open(json_filename))
+    test_df_dict = read_json_to_dataframes(summary_json)
 
     dict_len = len(test_df_dict.keys())
     test_df = test_df_dict["COPERNICUS/S2"]
