### 2020-07-27 alan-turing-institute/monitoring-ecosystem-resilience#367 @arbennett
Tutorial notebook errors
When running through the tutorial notebooks in the binder environment I ran into several issues which kept me from completing the tutorials.

## `tutorial_download_and_process_gee_images.ipynb`

I made it to the sixth code cell which is supposed to run:

```
urls = ee_prep_data(collection_dict, coords, date_range)
```

but I got the error:

```
HttpError                                 Traceback (most recent call last)
/srv/conda/envs/notebook/lib/python3.7/site-packages/ee/data.py in _execute_cloud_call(call, num_retries)
    337   try:
--> 338     return call.execute(num_retries=num_retries)
    339   except googleapiclient.errors.HttpError as e:

/srv/conda/envs/notebook/lib/python3.7/site-packages/googleapiclient/_helpers.py in positional_wrapper(*args, **kwargs)
    133                     logger.warning(message)
--> 134             return wrapped(*args, **kwargs)
    135

/srv/conda/envs/notebook/lib/python3.7/site-packages/googleapiclient/http.py in execute(self, http, num_retries)
    906         if resp.status >= 300:
--> 907             raise HttpError(resp, content, uri=self.uri)
    908         return self.postproc(resp, content)

HttpError: <HttpError 400 when requesting https://earthengine.googleapis.com/v1alpha/projects/earthengine-legacy/thumbnails?fields=name&alt=json returned "Total request size (34747888 bytes) must be less than or equal to 33554432 bytes.">

EEException                               Traceback (most recent call last)
<ipython-input-6-1cfaef990b71> in <module>
----> 1 urls = ee_prep_data(collection_dict, coords, date_range)

/srv/conda/envs/notebook/lib/python3.7/site-packages/pyveg/src/gee_interface.py in ee_prep_data(collection_dict, coords, date_range, region_size, scale, mask_cloud)
    208     for image in image_list:
    209         # get a URL from which we can download the resulting data
--> 210         url = image.getDownloadURL({"region": region, "scale": scale})
    211         url_list.append(url)
    212

/srv/conda/envs/notebook/lib/python3.7/site-packages/ee/image.py in getDownloadURL(self, params)
    437     request = params or {}
    438     request['image'] = self
--> 439     return data.makeDownloadUrl(data.getDownloadId(request))
    440
    441   def getThumbId(self, params):

/srv/conda/envs/notebook/lib/python3.7/site-packages/ee/data.py in getDownloadId(params)
    935     result = _execute_cloud_call(
    936         _cloud_api_resource.projects().thumbnails().create(
--> 937             parent=_get_projects_path(), fields='name', body=request))
    938     return {'docid': result['name'], 'token': ''}
    939

/srv/conda/envs/notebook/lib/python3.7/site-packages/ee/data.py in _execute_cloud_call(call, num_retries)
    338     return call.execute(num_retries=num_retries)
    339   except googleapiclient.errors.HttpError as e:
--> 340     raise _translate_cloud_exception(e)
    341
    342

EEException: Total request size (34747888 bytes) must be less than or equal to 33554432 bytes.
```

## `tutorial_analyse_gee_data.ipynb`

I made it to the fourth code cell which runs:
```
print(f"Reading results from '{os.path.abspath(json_summary_path)}'...")
dfs = variable_read_json_to_dataframe(json_summary_path)

print (dfs.keys())
```

But ran into the error:
```
Reading results from '/home/jovyan/notebooks/tutorials/results_summary_TigerBush_Niger.json'...

NameError                                 Traceback (most recent call last)
<ipython-input-4-c1085c1d4a26> in <module>
      1 print(f"Reading results from '{os.path.abspath(json_summary_path)}'...")
----> 2 dfs = variable_read_json_to_dataframe(json_summary_path)
      3
      4 print (dfs.keys())

NameError: name 'variable_read_json_to_dataframe' is not defined
```
