### 2020-07-27 alan-turing-institute/monitoring-ecosystem-resilience#368 @arbennett
Example configuration doesn't seem to run through
In reviewing this for JOSS I attempted to run through the example configuration, but think something may have gone wrong (possibly user error). I started by following the installation instructions which went smoothly. Then, I ran the example configuration to download the data:

```
pyveg_run_pipeline --config pyveg/configs/config_all.py
```

Which produced a directory with all of the downloaded data (`aztestoutput__2020-07-27_10-40-15`). I then tried to use this data to run the `pyveg_gee_analysis` script via:

```
pyveg_gee_analysis --input_dir aztestoutput__2020-07-27_10-40-15/gee_27.94_11.58_combine/
```

Which produced the following output (and error):

```
-----------------------------------
Running analyse_gee_data.py
-----------------------------------
Reading results from "/home/bzq/workspace/monitoring-ecosystem-resilience/aztestoutput__2020-07-27_10-40-15/gee_27.94_11.58_combine/results_summary.json"...

Preprocessing data...
---------------------
- Dropping vegetation outliers...
- Fill gaps in sub-image time series...
Traceback (most recent call last):
  File "/home/bzq/.conda/envs/veg/bin/pyveg_gee_analysis", line 8, in <module>
    sys.exit(main())
  File "/home/bzq/.conda/envs/veg/lib/python3.7/site-packages/pyveg/scripts/analyse_gee_data.py", line 284, in main
    analyse_gee_data(input_dir, spatial)
  File "/home/bzq/.conda/envs/veg/lib/python3.7/site-packages/pyveg/scripts/analyse_gee_data.py", line 204, in analyse_gee_data
    ts_dirname, dfs = preprocess_data(input_dir, n_smooth=4, resample=False, period='MS')
  File "/home/bzq/.conda/envs/veg/lib/python3.7/site-packages/pyveg/src/analysis_preprocessing.py", line 806, in preprocess_data
    dfs = fill_veg_gaps(dfs, missing)
  File "/home/bzq/.conda/envs/veg/lib/python3.7/site-packages/pyveg/src/analysis_preprocessing.py", line 562, in fill_veg_gaps
    df_ = df_.append(missing_rows, ignore_index=True).sort_values(by='date')
  File "/home/bzq/.local/lib/python3.7/site-packages/pandas/core/frame.py", line 6199, in append
    elif isinstance(other, list) and not isinstance(other[0], DataFrame):
IndexError: list index out of range

```

I believe this is related to #350
