### 2020-07-27 alan-turing-institute/monitoring-ecosystem-resilience#369 @arbennett
Documenting core functionality
In reviewing this for JOSS I found it difficult to understand how the structure of the software relates to the core functionalities. It seems several command line tools have been developed, but also are able to be operated via an API from within a python session as well. I think it would be very helpful to develop documentation outlining the correspondence between these two interfaces, as well as some general API documentation browse-able. I see the code is very nicely commented, so simply running it through something like sphinx autodoc would be really nice. Additionally, a mapping between names/modules in code and the pyveg flow diagram would make it very clear how to run each aspect of the workflow individually.

Broken out into a a checkbox list, I would find it very helpful to have the following:
- [x] - overview of the command line tools which are installed
- [ ] - overview of the high-level API that is used in the tutorial notebooks
- [x] - API autodocumentation generated from the code
- [x] - Correspondence between pyveg flow diagram and variable/class/module names in the codebase
