### 2020-07-27 alan-turing-institute/monitoring-ecosystem-resilience#370 @arbennett
Document configuration options
In a similar, yet slightly different, vein as #369 I think it would be very helpful to have documentation on the kinds of configuration that is allowable in pyveg. For instance, it is unclear what the available options are for `output_location_type` or `collections_to_use`. A simple listing of all of the user-modifiable options that go in the configuration file would clear this up.
