### 2020-07-27 alan-turing-institute/monitoring-ecosystem-resilience#371 @arbennett
Allow user configuration of coordinates
Currently the `coordinates.py` file is hardcoded to have 33 locations available for analysis, but as a community tool I think this should be user-configurable without having to dive into the source code. It seems there's already a pretty consistent definition where each entry requires:

* `continent`
* `country`
* `type`
* `latitude`
* `longitude`

So, it seems a simple csv file should be allowable to use as input (or even command line flags for these).
