### 2020-07-27 #2483 @arbennett
:wave: Hi @samvanstroud - thanks for developing pyveg, I think it sounds like a very nice tool for processing GEE vegetation data. However, in reviewing this I came across a few difficulties that I think need to be addressed before accepting this for publication. I have opened a few issues on the `monitoring-ecosystem-resilience` for organization. Here is a list:

- [] - [Example configuration doesn't seem to run through](https://github.com/alan-turing-institute/monitoring-ecosystem-resilience/issues/368)
- [] - [Tutorial notebook errors](https://github.com/alan-turing-institute/monitoring-ecosystem-resilience/issues/367)
- [] - [Documenting core functionality](https://github.com/alan-turing-institute/monitoring-ecosystem-resilience/issues/369)
- [] - [Documenting configuration](https://github.com/alan-turing-institute/monitoring-ecosystem-resilience/issues/370)
- [] - [Allow user configuration of coordinates](https://github.com/alan-turing-institute/monitoring-ecosystem-resilience/issues/371)
