### 2020-07-28 d4c35e5
add name of config file to attributes of pipeline
