### 2020-07-28 1312b91
add metadata dict to results_summary.json
