### 2020-07-30 0b9ea58
first iteration of Azure documentation
