+++ pyveg/UsingAzure.md
@@ -1,0 +1,103 @@
+# Running `pyveg` on the cloud - Microsoft Azure
+
+It is possible to make use of Azure cloud infrastructur in two ways when running pyveg:
+* Using Azure blob storage to store downloaded images and results of the image processing.
+* Using Azure batch to parallelize the running of the image processing.  This can vastly speed up the running time of your job.
+
+In order to do these, you will need the following:
+* An Azure account and an active subscription.
+* An Azure "Blob Storage Account" - follow instructions on https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-create-account-block-blob?tabs=azure-portal
+* An Azure "Batch Account" - see https://docs.microsoft.com/en-us/azure/batch/batch-technical-overview for a description.  When setting up the batch account, it will ask you to link it to a Storage Account, so use the one above (and ensure that you select the same "Region" for both.
+* You will probably need to increase the "quota" on your Batch Account - I believe that by default the quota for different types of VM are all set to zero.  There are instructions on how to do this at: https://docs.microsoft.com/en-us/azure/batch/batch-quota-limit - for our workflow we are using 100 dedicated cores of A1v2 VMs.
+
+## Setting up `azure_config.py`
+
+In the `pyveg/` directory there is a file `azure_config_template.py`.  Copy this to `azure_config.py` (in the same directory, and then start filling in the various fields.  The necessary info can be found in the Azure portal [https://portal.azure.com/#home] - perhaps the easiest way is to navigate via the "Subscriptions" icon at the top of the portal, then find the "Resources" (i.e. the Storage Account and the Batch Account).
+* For the Storage Account, look on the left sidebar under "Settings" for the "Access keys" - then copy/paste one of the keys into the relevant field in `azure_config.py`.
+* For the Batch Account, similarly there is a "Keys" icon under "Settings" which will lead to the relevant info.
+For the "batch_pool_id" you can put any name you like - if there will be multiple people using the same batch and storage accounts, you might want to use your initials or something to identify you (and in this case, you should be careful that the sum of everyone's "node_count"s don't exceed the quota for the batch account.
+
+Once you have populated the fields in `azure_config.py`, then do
+```
+pip install .
+```
+from the main `monitoring-ecosystem-resilience` directory.
+
+## config settings to use Azure storage
+
+We recommend that you create a new config file in the `pyveg/configs/` directory for every location/collection that you run.  You can use [pyveg/configs/test_Sentinel2_azure.py] as an example (you will want to increase the "date_range", and "n_sub_images" for the `NetworkCentralityCalculator` before running production though).
+
+The key setting to tell the job to use Azure blob storage is:
+```
+output_location_type = "azure"
+```
+
+### config settings to use Azure batch
+
+Note that using Azure storage as detailed above is a prerequisite for using Azure batch.
+Again there is an example template [pyveg/configs/test_Sentinel2_batch.py] that you can copy and modify with your own coordinates, date range, collection etc.
+
+Here, the important settings that enable the time-consuming parts of the pipeline to use Azure batch are in the `special_config` dictionary at the bottom of the file:
+```
+special_config = {
+    "VegetationImageProcessor": {"run_mode": "batch"},
+    "NetworkCentralityCalculator": {
+        "n_sub_images": 10,
+        "n_threads": 1,
+        "run_mode": "batch",
+    },
+    "NDVICalculator": {"run_mode": "batch"},
+}
+```
[…]
