### 2020-07-31 0d13368
add endpoint to download from azure
