### 2020-07-31 629e115
add test for metadata in combiner modules test
