### 2020-07-31 681fcf4
fix typo in test for metadata in combiner modules test
