### 2020-08-04 9b0b811
function to crease summary stats on the time series (#363)
