### 2020-08-06 d3c3013
adding comments to the fuction (#363)
