### 2020-08-06 f9a14e3
removing old commented code (#380)
