### 2020-08-06 fdb5429
allow Sequence configuration to be overriden via special_config in config file
