### 2020-08-06 #2483 @crangelsmith
👋Hi @arbennett , thank you for your comments and issues, they are very relevant and some where already in our to-do list! @nbarlowATI  and I will be working through these issues in the next couple of weeks and keep you updated.
