### 2020-08-07 a11cc78
steps towards masking some subimages
