+++ pyveg/src/analysis_preprocessing.py
@@ -16,61 +16,111 @@ from statsmodels.nonparametric.smoothers_lowess import lowess
 
 from pyveg.src.data_analysis_utils import write_to_json
 
+try:
+    from pyveg.src import azure_utils
+except:
+    print("Unable to import azure_utils")
 
-def read_json_to_dataframes(data):
+def read_results_summary(input_location, input_location_type):
     """
-    convert json data to a dict of DataFrame.
+    Read the results_summary.json, either from local storage or from Azure blob storage.
+
+    Parameters
+    ==========
+    input_location: str, directory or container with results_summary.json in
+    input_location_type: str: 'local' or 'azure'
+
+    Returns
+    =======
+    json_data: dict, the contents of results_summary.json
+    """
+    json_name = "results_summary.json"
+    if input_location_type == "local":
+        json_filepath = os.path.join(input_location, json_name)
+        if not os.path.exists(json_filepath):
+            raise FileNotFoundError("Unable to find {}".format(json_filepath))
+        json_data = json.load(open(json_filepath))
+        return json_data
+    elif input_location_type == "azure":
+        subdirs = azure_utils.list_directory(input_location, input_location)
+        print("Found subdirs {}".format(subdirs))
+        for subdir in subdirs:
+            print("looking at subdir {}".format(subdir))
+            if "combine" in subdir:
+                files = azure_utils.list_directory(input_location+"/"+subdir,
+                                                   input_location)
+                if json_name in files:
+                    return azure_utils.read_json(input_location+"/"+subdir+"/"+json_name,
+                                                 input_location)
+                else:
+                    raise RuntimeError("No {} found in {}".format(json_name, subdir))
+        return {}
+    else:
+        raise RuntimeError("input_location_type needs to be either 'local' or 'azure'")
+
+
+def read_results_summary(input_location, input_location_type):
+    """
+    Read the results_summary.json, either from local storage or from Azure blob storage.
+
+    Parameters
+    ==========
+    input_location: str, directory or container with results_summary.json in
+    input_location_type: str: 'local' or 'azure'
+
+    Returns
+    =======
+    json_data: dict, the contents of results_summary.json
+    """
+    json_name = "results_summary.json"
+    if input_location_type == "local":
+        json_filepath = os.path.join(input_location, json_name)
+        if not os.path.exists(json_filepath):
+            raise FileNotFoundError("Unable to find {}".format(json_filepath))
+        json_data = json.load(open(json_filepath))
+        return json_data
+    elif input_location_type == "azure":
+        subdirs = azure_utils.list_directory(input_location, input_location)
+        print("Found subdirs {}".format(subdirs))
+        for subdir in subdirs:
+            print("looking at subdir {}".format(subdir))
+            if "combine" in subdir:
+                files = azure_utils.list_directory(input_location+"/"+subdir,
+                                                   input_location)
+                if json_name in files:
+                    return azure_utils.read_json(input_location+"/"+subdir+"/"+json_name,
+                                                 input_location)
+                else:
+                    raise RuntimeError("No {} found in {}".format(json_name, subdir))
+        return {}
+    else:
+        raise RuntimeError("input_location_type needs to be either 'local' or 'azure'")
+
+
+
+def read_label_json(input_label_json):
+    """
+    Read JSON output of ImageLabeller, and return a list of coords to be masked out
 
     Parameters
     ----------
-    data : dict, json data output from run_pyveg_pipeline
+    input_label_json : str, path to json file
 
     Returns
     ----------
-    dict
-        A dict of the saved results in a DataFrame format. Keys are
-        names of collections and the values are DataFrame of results
-        for that collection.
[…]
