### 2020-08-07 34c8d31
changing notebook to new function names and make sure plots are shown
