### 2020-08-10 f39fc21
use '/' in directory paths for Azure, even when running on Windows
