### 2020-08-10 c3a0d14
updating text in notebook  (#367)
