### 2020-08-11 d01e24b
Added R functions to calculate 3 decay rates
