### 2020-08-11 #2483 @usethedata
@crangelsmith -- Thanks for the comment.  I'm thinking I'll do the second review, as I'd like to dig into this work more carefully than I have so far, as it relates to some of what I do in my day job.
