### 2020-08-12 8cef442
making sure in the analysis that data is a time series (at least one date) (#368)
