### 2020-08-12 69d6f7a
fixing report for the cases where there is not time series (#368)
