### 2020-08-13 21e6e3b
WIP towards reading mask file for analysis preprocessing
