+++ pyveg/src/analysis_preprocessing.py
@@ -21,22 +21,26 @@ try:
 except:
     print("Unable to import azure_utils")
 
-def read_results_summary(input_location, input_location_type):
+
+def read_results_summary(input_location,
+                         input_filename="results_summary.json",
+                         input_location_type="local"):
     """
     Read the results_summary.json, either from local storage or from Azure blob storage.
 
     Parameters
     ==========
     input_location: str, directory or container with results_summary.json in
+    input_filename: str, name of json file, default is "results_summary.json"
     input_location_type: str: 'local' or 'azure'
 
     Returns
     =======
     json_data: dict, the contents of results_summary.json
     """
-    json_name = "results_summary.json"
+
     if input_location_type == "local":
-        json_filepath = os.path.join(input_location, json_name)
+        json_filepath = os.path.join(input_location, input_filename)
         if not os.path.exists(json_filepath):
             raise FileNotFoundError("Unable to find {}".format(json_filepath))
         json_data = json.load(open(json_filepath))
@@ -49,49 +53,11 @@ def read_results_summary(input_location, input_location_type):
             if "combine" in subdir:
                 files = azure_utils.list_directory(input_location+"/"+subdir,
                                                    input_location)
-                if json_name in files:
-                    return azure_utils.read_json(input_location+"/"+subdir+"/"+json_name,
+                if input_filename in files:
+                    return azure_utils.read_json(input_location+"/"+subdir+"/"+input_filename,
                                                  input_location)
                 else:
-                    raise RuntimeError("No {} found in {}".format(json_name, subdir))
-        return {}
-    else:
-        raise RuntimeError("input_location_type needs to be either 'local' or 'azure'")
-
-
-def read_results_summary(input_location, input_location_type):
-    """
-    Read the results_summary.json, either from local storage or from Azure blob storage.
-
-    Parameters
-    ==========
-    input_location: str, directory or container with results_summary.json in
-    input_location_type: str: 'local' or 'azure'
-
-    Returns
-    =======
-    json_data: dict, the contents of results_summary.json
-    """
-    json_name = "results_summary.json"
-    if input_location_type == "local":
-        json_filepath = os.path.join(input_location, json_name)
-        if not os.path.exists(json_filepath):
-            raise FileNotFoundError("Unable to find {}".format(json_filepath))
-        json_data = json.load(open(json_filepath))
-        return json_data
-    elif input_location_type == "azure":
-        subdirs = azure_utils.list_directory(input_location, input_location)
-        print("Found subdirs {}".format(subdirs))
-        for subdir in subdirs:
-            print("looking at subdir {}".format(subdir))
-            if "combine" in subdir:
-                files = azure_utils.list_directory(input_location+"/"+subdir,
-                                                   input_location)
-                if json_name in files:
-                    return azure_utils.read_json(input_location+"/"+subdir+"/"+json_name,
-                                                 input_location)
-                else:
-                    raise RuntimeError("No {} found in {}".format(json_name, subdir))
+                    raise RuntimeError("No {} found in {}".format(input_filename, subdir))
         return {}
     else:
         raise RuntimeError("input_location_type needs to be either 'local' or 'azure'")
@@ -119,8 +85,65 @@ def read_label_json(input_label_json):
         return []
     df = pd.DataFrame.from_records(json.load(open(input_label_json)))
     top_labels = df.groupby(["longitude","latitude"]).category.max().to_dict()
-    coords_to_mask = [k for k,v in top_labels.values() if v=="Not patterned vegetation"]
+    coords_to_mask = [k for k,v in top_labels.items() if v=="Not patterned vegetation"]
     return coords_to_mask
+
+
[…]
