### 2020-08-13 8d9c25f
change os.path.join to '/'.join when downloading from azure
