### 2020-08-14 bfce109
convert chris's r functions to python
started translating the r code to python. not near done but wanted to push it now in case my computer crashes again :)
