+++ rveg/R/generatePattern.py
@@ -1,0 +1,154 @@
+### adapted from exerciseflat.m matlab code
+#library(ggplot2)
+#library(gganimate)
+#library(reshape2)
+#library(jsonlite)
+#library(dplyr)
+
+#' Load configuration from a JSON file
+#' @param configFile Name of the JSON file containing configuration parameters
+#' @return config an environment containing the configured parameters.
+#' @export
+def loadConfig(configFile=file_path("..", "testdata", "patternConfig.json")):
+    config = read_json(configFile)
+    return config
+
+#' Given a 2D array of floats, return a 2D array (same dimensions) of 1s and 0s.
+#' By default, take the threshold to be halfway between min and max values.
+#' @param pattern Matrix of floats representing biomass in each square of a grid
+#' @param threshold Optionally specify a threshold above which we will put a 1 in the output matrix.
+def binarizePattern(pattern, threshold=None):
+    if is_null(threshold):
+        threshold = (max(pattern) + min(pattern)) / 2
+    binaryPattern = (pattern > threshold) * 1
+    return binaryPattern
+
+#' Load a CSV file containing a 2D array of 1s and zeros, representing a starting pattern.
+#' If no file is given, start with all zeros.
+#' @param patternFile filename of CSV file containing starting pattern of 1s and 0s.
+#' @param m length of each side of the matrix.
+#' @return pattern m*m matrix of 1s and 0s.
+#' @export
+def getStartingPattern(patternFile=None, m=50):
+    pattern = matrix(0, m, m)
+    if ~  is_null(patternFile):
+        pattern = as_matrix(read_csv(patternFile, header=False))
+        pattern = pattern[range(1, m), range(1, m)]
+    return pattern
+
+#' Use the model to evolve the starting pattern through several timesteps
+#' keeping track of surface and soil water.  Return the final pattern.
+#' @param configFile JSON configuration file
+#' @param startingPatternFilename CSV file containing initial pattern (if not provided, will be all 0s).
+#' @return popP Final pattern - a 2D array.
+#' @export
+generatePattern <- function(configFile=file.path("..","testdata","patternGenConfig.json"),
+                            startingPatternFilename=NULL) {
+    print(getwd())
+    config <- loadConfig(configFile)
+    ## System discretisation
+    DeltaX <- config$DeltaX # (m)
+    DeltaY <- config$DeltaY # (m)
+    ## Diffusion constants for plants, soil water and surface water (see remark above)
+    DifP <- config$DifP   # (m2.d-1)
+    DifW <- config$DifW   # (m2.d-1)
+    DifO <- config$DifO   # (m2.d-1)
+    ## Initial fraction of grid cells with bare ground
+    frac <- config$frac  # (-)
+
+    ## Parameter values
+    R       <-   config$R      # Rainfall (mm.d-1)
+    alpha   <-   config$alpha  # Proportion of surface water available for infiltration (d-1)
+    W0      <-   config$W0     # Bare soil infiltration (-)
+    beta    <-   config$beta   # Plant loss rate due to grazing (d-1)
[…]
