+++ rveg/R/generatePattern.py
@@ -5,6 +5,9 @@
 #library(jsonlite)
 #library(dplyr)
 
+import pandas as pd
+import numpy as np
+
 #' Load configuration from a JSON file
 #' @param configFile Name of the JSON file containing configuration parameters
 #' @return config an environment containing the configured parameters.
@@ -42,82 +45,68 @@ def getStartingPattern(patternFile=None, m=50):
 #' @param startingPatternFilename CSV file containing initial pattern (if not provided, will be all 0s).
 #' @return popP Final pattern - a 2D array.
 #' @export
-generatePattern <- function(configFile=file.path("..","testdata","patternGenConfig.json"),
-                            startingPatternFilename=NULL) {
+def generatePattern(configFile=file.path("..","testdata","patternGenConfig.json"), startingPatternFilename=NULL):
     print(getwd())
-    config <- loadConfig(configFile)
+    config = loadConfig(configFile)
     ## System discretisation
-    DeltaX <- config$DeltaX # (m)
-    DeltaY <- config$DeltaY # (m)
+    DeltaX = config$DeltaX # (m)
+    DeltaY = config$DeltaY # (m)
     ## Diffusion constants for plants, soil water and surface water (see remark above)
-    DifP <- config$DifP   # (m2.d-1)
-    DifW <- config$DifW   # (m2.d-1)
-    DifO <- config$DifO   # (m2.d-1)
+    DifP = config$DifP   # (m2.d-1)
+    DifW = config$DifW   # (m2.d-1)
+    DifO = config$DifO   # (m2.d-1)
     ## Initial fraction of grid cells with bare ground
-    frac <- config$frac  # (-)
-
+    frac = config$frac  # (-)
     ## Parameter values
-    R       <-   config$R      # Rainfall (mm.d-1)
-    alpha   <-   config$alpha  # Proportion of surface water available for infiltration (d-1)
-    W0      <-   config$W0     # Bare soil infiltration (-)
-    beta    <-   config$beta   # Plant loss rate due to grazing (d-1)
-    rw      <-   config$rw     # Soil water loss rate due to seepage and evaporation (d-1)
-    c       <-   config$c      # Plant uptake constant (g.mm-1.m-2)
-    gmax    <-   config$gmax   # Plant growth constant (mm.g-1.m-2.d-1)
-    d       <-   config$d      # Plant senescence rate (d-1)
-    k1      <-   config$k1     # Half saturation constant for plant uptake and growth (mm)
-    k2      <-   config$k2     # Half saturation constant for water infiltration (g.m-2)
-
+    R = config$R      # Rainfall (mm.d-1)
+    alpha = config$alpha  # Proportion of surface water available for infiltration (d-1)
+    W0 = config$W0     # Bare soil infiltration (-)
+    beta = config$beta   # Plant loss rate due to grazing (d-1)
+    rw = config$rw     # Soil water loss rate due to seepage and evaporation (d-1)
+    c = config$c      # Plant uptake constant (g.mm-1.m-2)
+    gmax = config$gmax   # Plant growth constant (mm.g-1.m-2.d-1)
+    d = config$d      # Plant senescence rate (d-1)
+    k1 = config$k1     # Half saturation constant for plant uptake and growth (mm)
+    k2 = config$k2     # Half saturation constant for water infiltration (g.m-2)
     ## Number of grid cells
-    m <- config$m
-    NX <- m
-    NY <- m
-
+    m = config$m
+    NX = m
+    NY = m
     ## Timesteps
[…]
