+++ rveg/R/EC_calc.py
@@ -6,6 +6,9 @@
 #       email: c.a.boulton@exeter.ac.uk
 #
 
+import pandas as pd
+import numpy as np
+
 #' Function to find the coordinates of pixels neighbouring a pixel at (x,y) in an m*n array.
 #' Will return an array of eight coordinates (if including diagonal neighbours) or four if not.
 #' @param x The x-coordinate of the pixel of interest
@@ -15,27 +18,26 @@
 #' @param diag Include diagonal neighbours if set to TRUE
 #' @return Array of neighbouring pixel's x-y coordinates.
 #' @export
-get_neighbours <- function(x,y,m,n,diag=TRUE) {
+def get_neighbours(x,y,m,n,diag=TRUE):
     #find indices of surrounding neighbours of point x,y in an m x n array
     #diagonal neighbours also returned as default
     #indices which lie outside of the boundaries are removed
-    if (diag == TRUE) {
-        neighbours <- array(NA, dim=c(8,2))
-        neighbours[1,] <- c(x-1,y+1)
-        neighbours[2,] <- c(x,y+1)
-        neighbours[3,] <- c(x+1,y+1)
-        neighbours[4,] <- c(x-1,y)
-        neighbours[5,] <- c(x+1,y)
-        neighbours[6,] <- c(x-1,y-1)
-        neighbours[7,] <- c(x,y-1)
-        neighbours[8,] <- c(x+1,y-1)
-    } else {
-        neighbours <- array(NA, dim=c(4,2))
-        neighbours[1,] <- c(x,y+1)
-        neighbours[2,] <- c(x-1,y)
-        neighbours[3,] <- c(x+1,y)
-        neighbours[4,] <- c(x,y-1)
-    }
+    if diag == TRUE:
+        neighbours = array(NA, dim=c(8,2))
+        neighbours[1,] = c(x-1,y+1)
+        neighbours[2,] = c(x,y+1)
+        neighbours[3,] = c(x+1,y+1)
+        neighbours[4,] = c(x-1,y)
+        neighbours[5,] = c(x+1,y)
+        neighbours[6,] = c(x-1,y-1)
+        neighbours[7,] = c(x,y-1)
+        neighbours[8,] = c(x+1,y-1)
+    else:
+        neighbours = array(NA, dim=c(4,2))
+        neighbours[1,] = c(x,y+1)
+        neighbours[2,] = c(x-1,y)
+        neighbours[3,] = c(x+1,y)
+        neighbours[4,] = c(x,y-1)
     if (length(which(neighbours[,1] > m |
                      neighbours[,1] < 1 |
                      neighbours[,2] > n |
@@ -45,8 +47,7 @@ get_neighbours <- function(x,y,m,n,diag=TRUE) {
                                         neighbours[,2] > n |
                                         neighbours[,2] < 1),]
     }
-    return(neighbours)
-}
+    return neighbours
 #
 
 #' Function that creates a 2-column array of all the edges where neighbour points are both equal to 1
@@ -115,23 +116,17 @@ def invert_pattern(pattern):
 #' @param pattern 2D array of 1s and zeros corresponding to a binary image.
 #' @param inc size of increment in percentage of pixels (ordered by subgraph centrality) in feature vector.
 #' @return EC_vec list of Euler Characteristic values (feature vector).
-calc_EC <- function(pattern, inc=5) {
-
-    require('igraph')
-    graph_edges <- make_edges(pattern)
-    graph <- igraph::graph_from_data_frame(graph_edges, directed = TRUE, vertices = NULL)
[…]
