### 2020-08-14 4ceb15f
fix EWS analysis for cases where there are only nan (#368)
