### 2020-08-14 359cd9e
do no make EWS analysis in short time series (less than 12 points) (#368)
