### 2020-08-14 f8c8447
make sure reporting does not include EWS for short time series (less than 12 points) (#368)
