### 2020-08-14 3ed50d6
remove seasonal cycle removal for times series of less than 24 months (#368)
