### 2020-08-14 d521ff4
example download config has at least 3 months (#368)
