### 2020-08-14 1270fa6
at least 3 months for time series analysis (#368)
