### 2020-08-14 eb46fab
adding metadata to the summary stats function (#405)
