### 2020-08-14 8e59b5f
adding metadata and kendall tau values to the summary stats function (#405)
