### 2020-08-14 8968061
adding info of rolling window into variables (#405)
