### 2020-08-14 91b7d90
Added Crystal Ball function
Function takes in time series (such as mean_offset50) and outputs parameters estimated for Crystal Ball pdf using nonlinear least squares
