### 2020-08-14 259318e
adding initial script for analysing summary data (#407)
