### 2020-08-17 f2a2b98
Delete EC_calc.py
