+++ rveg/R/EC_calc.py
@@ -1,132 +1,0 @@
-#       Functions to calculate the feature vecture of patterned vegetation as in Mander et al. 2017
-#       Patterns should supplied as a binary [0,1] m x n array
-#
-#       Author: Chris A. Boulton
-#       Date: 21th November 2019
-#       email: c.a.boulton@exeter.ac.uk
-#
-
-import pandas as pd
-import numpy as np
-
-#' Function to find the coordinates of pixels neighbouring a pixel at (x,y) in an m*n array.
-#' Will return an array of eight coordinates (if including diagonal neighbours) or four if not.
-#' @param x The x-coordinate of the pixel of interest
-#' @param y The y-coordinate of the pixel of interest
-#' @param m The size of the array in x-direction.
-#' @param n The size of the array in y-direction.
-#' @param diag Include diagonal neighbours if set to TRUE
-#' @return Array of neighbouring pixel's x-y coordinates.
-#' @export
-def get_neighbours(x,y,m,n,diag=TRUE):
-    #find indices of surrounding neighbours of point x,y in an m x n array
-    #diagonal neighbours also returned as default
-    #indices which lie outside of the boundaries are removed
-    if diag == TRUE:
-        neighbours = array(NA, dim=c(8,2))
-        neighbours[1,] = c(x-1,y+1)
-        neighbours[2,] = c(x,y+1)
-        neighbours[3,] = c(x+1,y+1)
-        neighbours[4,] = c(x-1,y)
-        neighbours[5,] = c(x+1,y)
-        neighbours[6,] = c(x-1,y-1)
-        neighbours[7,] = c(x,y-1)
-        neighbours[8,] = c(x+1,y-1)
-    else:
-        neighbours = array(NA, dim=c(4,2))
-        neighbours[1,] = c(x,y+1)
-        neighbours[2,] = c(x-1,y)
-        neighbours[3,] = c(x+1,y)
-        neighbours[4,] = c(x,y-1)
-    if (length(which(neighbours[,1] > m |
-                     neighbours[,1] < 1 |
-                     neighbours[,2] > n |
-                     neighbours[,2] < 1)) > 0) {
-        neighbours <- neighbours[-which(neighbours[,1] > m |
-                                        neighbours[,1] < 1 |
-                                        neighbours[,2] > n |
-                                        neighbours[,2] < 1),]
-    }
-    return neighbours
-#
-
-#' Function that creates a 2-column array of all the edges where neighbour points are both equal to 1
-#' @param pattern A 2D binary array
-#' @return edges A 2-column array of all edges between signal pixels
-#' @export
-def make_edges():
-    m = dim(pattern)[[1]]
-    n = dim(pattern)[[2]]
-    # set this up as an array to add to the bottom of and then delete the top
-    # two rows at the end
-    edges = array(0, dim=make_tuple(2, 2)) for x in range(1, m):
-    for y in range(1, n):
-        if pattern[x, y] == 1:
-            neigh = NA
-            neigh = get_neighbours(x, y, m, n, diag=True)
-            for i in range(1, dim(neigh)[1]):
-                if pattern[neigh[i, 1], neigh[i, 2]] == 1:
-                    # trial and error has been used to work out which indices
-                    # to record
-                    edges = rbind(edges, make_tuple(
-                        (y - 1) * m + x, (neigh[i, 2] - 1) * m + neigh[i, 1]))
-    edges = edges[- make_tuple(range(1, 2))
-    return edges
-
-
-#' Read values from a CSV file into a matrix.
-#' @param filename Full path to csv file
-#' @return matrix containing 1s and 0s.
-#' @export
-def read_from_csv(filename):
-    df = read_csv(filename, header=False)
[…]
