### 2020-08-17 c7dcba4
Delete generatePattern.py
