### 2020-08-17 8573e0d
Update and rename rveg/crystal_ball_function.R to rveg/R/crystal_ball_function.R
Fixed the mean_annual_cycle function
