### 2020-08-17 6160040
Update decay_rate_functions.R
Fixed mean_annual_cycle function
