### 2020-08-17 27f6f2e
translated decay_rate_functions to python
i translated the decay_rate_functions code from R to python. Should run. Havent gotten to cross check results yet
