+++ rveg/R/decay_rate_functions.py
@@ -3,6 +3,7 @@
 #	Looks at the mean annual time series to calculate rate on (from max to min)
 #
 #	Author: Chris A. Boulton
+#   Translated from R to python by Jesse F. Abrams
 #	Date: 11th August 2020
 #	email: c.a.boulton@exeter.ac.uk
 #
@@ -15,13 +16,14 @@ def mean_annual_ts(x, resolution=12):
     # resolution is the number of time points per year in the time series
     # infills missing data by linear interpolation
     # note that the first or last value will not be infilled if missing
-    missing_inds = which(is_na(x))
-    if length(missing_inds) > 0:
-        for i in range(1, length(missing_inds)):
-            x[i] = mean(make_tuple(x[i - 1], x[i + 1]))
-    mean_cycle = rep(NA, resolution)
-    for i in range(1, resolution):
-        mean_cycle[i] = mean(x[seq(i, length(x), resolution)], na_rm=T)
+    missing_inds = np.where(np.isnan(x))[0]
+    if len(missing_inds) > 0:
+        for i in range(len(missing_inds)):
+            print(i)
+            x[missing_inds[i]] = np.mean([x[missing_inds[i] - 1], x[missing_inds[i] + 1]])
+    mean_cycle = np.repeat(np.nan, resolution, axis=0)
+    for i in range(resolution):
+        mean_cycle[i] = np.nanmean(x[np.linspace(start=i, stop=len(x)-1, num=resolution,dtype=int)])
     return mean_cycle
 
 
@@ -30,49 +32,49 @@ def decay_rate(x, resolution=12, method='basic'):
     # then calculates the decay rate between the max and min values
     # choice of 'basic' or 'adjusted' for method:
     #   basic - uses the raw values
-    # adjusted - substracts the minimum value and adds 1 to time series before
+    # adjusted - subtracts the minimum value and adds 1 to time series before
     # calculation
     annual_cycle = mean_annual_ts(x, resolution)
     if method == 'basic':
         ts = annual_cycle
     elif method == 'adjusted':
-        ts = annual_cycle - min(annual_cycle) + 1
+        ts = annual_cycle - np.min(annual_cycle) + 1
     else:
-        ts = NA  # causes fail if method is not specified properly
-    max_ind = which_max(ts)
-    min_ind = which_min(ts)
+        ts = np.nan  # causes fail if method is not specified properly
+    max_ind = np.where(ts==np.max(ts))[0][0]
+    min_ind = np.where(ts==np.min(ts))[0][0]
     if min_ind < max_ind:
         # this ensures the length of time for decay is correct below
         min_ind = min_ind + resolution
-    dr = log(min(ts) / max(ts)) / (min_ind - max_ind)
+    dr = np.log(np.min(ts) / np.max(ts)) / (min_ind - max_ind)
     return dr
 
 
 def exp_model_fit(x, resolution=12, method='basic'):
-	#takes a time series as an input and uses mean_annual_ts()
-	#then fits a linear regression model on the time series from max to min value
-	#choice of 'basic' or 'adjusted' for method:
-	#	basic - uses the raw values
-	#	adjusted - substracts the minimum value and adds 1 to time series before calculation
-	#####	NB: offset50 values are negative by default and so have to use adjusted method #####
-	annual_cycle = mean_annual_ts(x, resolution)
-	if method == 'basic':
-		ts = annual_cycle
-	elif method == 'adjusted':
-		ts = annual_cycle - min(annual_cycle) + 1
-	else:
-		ts = NA		#causes fail if method is not specified properly
-	max_ind = which.max(ts)
-	min_ind = which.min(ts)
[…]
