### 2020-08-17 c7a1eea
converting crystall ball function to python
