### 2020-08-17 45bfc77
further progress on python translation
