### 2020-08-17 ee457a2
Update decay_rate_functions.py
Changed lines 25 to chose the correct indices to calculate the mean monthly values and line 75 to swap the inputs of the polyfit function around.
