### 2020-08-17 3192a49
Slightly updated version of Crystal Ball function in Python
The optimisation does not appear to work and the mean_annual_ts function does not match up with what is the R functions
