### 2020-08-17 38f886b
adding scatterplots to summary (#407)
