### 2020-08-18 6cf108e
Update decay_rate_functions.R
Fixed mean_annual_ts function to chose correct missing_inds
