### 2020-08-18 4acaf06
Update decay_rate_functions.py
Changed len(x)-1 for len(x) on line 25
