### 2020-08-18 2d7bda8
Delete crystal_ball_function_v0.1.py
