### 2020-08-18 14b8214
Delete crystal_ball_function.py
