### 2020-08-18 aad41c1
Added decay rate and Crystal Ball functions
