+++ pyveg/src/data_analysis_utils.py
@@ -19,9 +19,12 @@ import matplotlib.pyplot as plt
 import matplotlib.cm as cm
 
 from scipy.fftpack import fft
-from scipy.stats import sem, t
+from scipy.stats import sem, t, norm
 from statsmodels.tsa.seasonal import STL
 import ewstools
+
+import scipy
+import scipy.optimize as sco
 
 
 def convert_to_geopandas(df):
@@ -1151,3 +1154,263 @@ def early_warnings_null_hypothesis(
     kendall_tau_df = pd.concat([data_kendall_tau_df, surrogates_kendall_tau_df])
 
     return kendall_tau_df
+
+def mean_annual_ts(x, resolution=12):
+    """
+    Calculate mean annual time series from time series. Also fills in missing values
+    by linear interpolation. NB Fails if there is missing value at the start or end.
+
+    Parameters
+    ----------
+    x : Time series
+        Time series to calculate mean annual time series for
+    resolution : float
+        Number of values each year in a time series (12 is monthly for example)
+
+    Returns
+    ----------
+    ndarray
+        Array of length equal to resolution that is the mean annual time series
+
+    """
+    
+    missing_inds = np.where(np.isnan(x))[0]
+    if len(missing_inds) > 0:
+        for i in range(len(missing_inds)):
+            print(i)
+            x[missing_inds[i]] = np.mean([x[missing_inds[i] - 1], x[missing_inds[i] + 1]])
+            
+    mean_cycle = np.repeat(np.nan, resolution, axis=0)
+    for i in range(resolution):
+        mean_cycle[i] = np.nanmean(x[np.linspace(start=i, stop=len(x)-1, num=resolution,dtype=int)])
+    return mean_cycle
+
+def decay_rate(x, resolution=12, method='basic'):
+    """
+    Calculates the decay rate between the max and min values of a time series.
+
+    Parameters
+    ----------
+    x : time series
+        Time series to calculate decay rate on. mean_annual_ts is calculated
+        on this series within this function so raw time series is expected.
+    resolution : int
+        Number of values each year in a time series (12 is monthly for example)
+    method : 'basic' (default) or 'adjusted'
+        A choice on whether to calculate the decay rate on the mean annual 
+        time series calculated within the function or to adjust the time series
+        such that the min value is set to 1 by substracting the minimum
+        plus 1 of the mean annual time series (useful for offset50 values)
+
+    Returns
+    ----------
+    float
+        The decay rate value
+
+    """
+    
+    annual_cycle = mean_annual_ts(x, resolution)
+    
+    if method == 'basic':
+        ts = annual_cycle
+    elif method == 'adjusted':
+        ts = annual_cycle - np.min(annual_cycle) + 1
+    else:
+        ts = np.nan  # causes fail if method is not specified properly
+        
+    max_ind = np.where(ts==np.max(ts))[0][0]
+    min_ind = np.where(ts==np.min(ts))[0][0]
+    
+    if min_ind < max_ind:
+        # this ensures the length of time for decay is correct below
+        min_ind = min_ind + resolution
+        
+    dr = np.log(np.min(ts) / np.max(ts)) / (min_ind - max_ind)
+    return dr
+
+def exp_model_fit(x, resolution=12, method='basic'):
+    """
+    Fits an exponential model from the maximum to the minimum of the 
+    mean annual time series. A raw time series is expected as an input.
+
+    Parameters
+    ----------
+    x : time series
+        Time series to calculate decay rate on. mean_annual_ts is calculated
[…]
