### 2020-08-18 f6d6dc5
Delete crystal_ball_function.py
