### 2020-08-18 a1701dc
Delete decay_rate_functions.py
