### 2020-08-18 ebb77d6
Update data_analysis_utils.py
Indent fixes
