+++ pyveg/src/data_analysis_utils.py
@@ -1,7 +1,6 @@
 """
 Data analysis code including functions to read the .json results file,
 and functions analyse and plot the data.
-
 """
 
 import json
@@ -31,12 +30,10 @@ def convert_to_geopandas(df):
     """
     Given a pandas DatFrame with `lat` and `long` columns, convert
     to geopandas DataFrame.
-
     Parameters
     ----------
     df : DataFrame
         Pandas DatFrame with `lat` and `long` columns.
-
     Returns
     ----------
     geopandas DataFrame
@@ -51,20 +48,17 @@ def convert_to_geopandas(df):
 def calculate_ci(data, ci_level=0.99):
     """
     Calculate the confidence interval on the mean for a set of data.
-
     Parameters
     ----------
     data : Series
         Series of data to calculate the confidence interval of the mean.
     ci_level : float, optional
         Size of the confidence interval to calculate
-
     Returns
     ----------
     float
         Confidence interval value where the CI is [mu - h, mu + h],
         where mu is the mean.
-
     """
 
     # remove NaNs
@@ -82,7 +76,6 @@ def get_confidence_intervals(df, column, ci_level=0.99):
     """
     Calculate the confidence interval at each time point of a
     DataFrame containing data for a large image.
-
     Parameters
     ----------
     df : DataFrame
@@ -91,7 +84,6 @@ def get_confidence_intervals(df, column, ci_level=0.99):
         Name of the column to calculate the CI of.
     ci_level : float, optional
         Size of the confidence interval to calculate
-
     Returns
     ----------
     DataFrame
@@ -119,7 +111,6 @@ def get_confidence_intervals(df, column, ci_level=0.99):
 def create_lat_long_metric_figures(geodf, metric, output_dir):
     """
     From input data-frame with processed network metrics create 2D gird figure for each date available using Geopandas.
-
     Parameters
     ----------
     geodf:  GeoDataframe
@@ -128,10 +119,8 @@ def create_lat_long_metric_figures(geodf, metric, output_dir):
         Variable to plot
     output_dir: string
         Directory to save the figures
-
      Returns
     ----------
-
     """
 
     if {"date", metric}.issubset(geodf.columns):
@@ -168,20 +157,16 @@ def coarse_dataframe(geodf, side_square):
     """
     Coarse the granularity of a dataframe by grouping lat,long points
     that are close to each other in a square of L = size_square
-
     Parameters
     ----------
     geodf:  Dataframe
         Input dataframe
     side_square: integer
         Side of the square
-
-
     Returns
     ----------
      A  dataframe
         A coarser dataframe
-
     """
 
     # initialise the categories
@@ -208,8 +193,8 @@ def coarse_dataframe(geodf, side_square):
                 for j in range(side_square):
 
                     if (
-                        n + n_grids * i + j < n_grids * n_grids
-                        and data_df["category"].iloc[n + n_grids * i + j] == -1
+                            n + n_grids * i + j < n_grids * n_grids
+                            and data_df["category"].iloc[n + n_grids * i + j] == -1
                     ):
                         indexes.append(n + n_grids * i + j)
 
@@ -244,9 +229,7 @@ def coarse_dataframe(geodf, side_square):
 
 def network_figure(df, date, metric, vmin, vmax, output_dir):
     """
-
     Make 2D heatmap plot with network centrality measures
-
     Parameters
     ----------
     df:  Dataframe
@@ -261,7 +244,6 @@ def network_figure(df, date, metric, vmin, vmax, output_dir):
         Colorbar max values
     output_dir: string
         Directory where to save the plots
-
     Returns
     ----------
     """
@@ -318,11 +300,9 @@ def network_figure(df, date, metric, vmin, vmax, output_dir):
[…]
