### 2020-08-18 ec5a778
added barplot function
added a barplot function but still need to make it prettier
