### 2020-08-19 3b25edd
script to generate config files
