### 2020-08-19 0f72a45
remove commented out code
