### 2020-08-19 ffb7536
prettify printout of generate_config_file
