### 2020-08-20 0a425b4
apply mask when reading subimage data in analysis_preprocessing
