### 2020-08-20 7c74422
better container names for azure
