### 2020-09-01 486089b
Allow daily ERA5 collection, and long-term historical ERA5 data
