### 2020-09-02 5356716
allow input of coordinate ids to generate_config_file.py
