### 2020-09-02 f8b3026
fix misleading error message on downloader
