### 2020-09-03 b785bd1
moves towards using metadata in analysis
