+++ pyveg/src/plotting.py
@@ -137,11 +137,11 @@ def plot_time_series(df, output_dir):
             ax.plot(veg_xs_b, veg_means_b, label='Unsmoothed', linewidth=1, color='indigo', linestyle='dashed', alpha=0.2)
 
             # plot smoothed vegetation means and std
-            ax3.plot(veg_xs_b, veg_means_smooth_b, marker='o', markersize=7, 
-                    markeredgecolor=(0.8172, 0.7627, 0.9172), markeredgewidth=2, 
+            ax3.plot(veg_xs_b, veg_means_smooth_b, marker='o', markersize=7,
+                    markeredgecolor=(0.8172, 0.7627, 0.9172), markeredgewidth=2,
                     label='Smoothed', linewidth=2, color=color)
 
-            ax3.fill_between(veg_xs_b, veg_means_smooth_b - veg_stds_smooth_b, veg_means_smooth_b + veg_stds_smooth_b, 
+            ax3.fill_between(veg_xs_b, veg_means_smooth_b - veg_stds_smooth_b, veg_means_smooth_b + veg_stds_smooth_b,
                             facecolor='tab:purple', alpha=0.1, label='Std Dev')
 
             # add veg-veg correlation
@@ -216,12 +216,12 @@ def plot_ndvi_time_series(df, output_dir):
         # plot ndvi
         #ax.plot(veg_xs, ndvi_means, label='Unsmoothed', linewidth=1, color='dimgray', linestyle='dotted')
 
-        ax.plot(veg_xs, veg_means, marker='o', markersize=7, 
+        ax.plot(veg_xs, veg_means, marker='o', markersize=7,
                 markeredgecolor=(0.9172, 0.9627, 0.9172), markeredgewidth=2,
                 label='Smoothed', linewidth=2, color='green')
 
         if any([col_name.replace('mean', 'std') == c for c in df.columns]):
-            ax.fill_between(veg_xs, veg_means - veg_std, veg_means + veg_std, 
+            ax.fill_between(veg_xs, veg_means - veg_std, veg_means + veg_std,
                             facecolor='green', alpha=0.1, label='Std Dev')
 
         # plot precipitation if availible
@@ -270,7 +270,7 @@ def plot_ndvi_time_series(df, output_dir):
             print(f'Plotting {veg_prefix} NDVI time series.')
             make_plot(df, veg_prefix, column, output_dir)
 
-    
+
 def plot_autocorrelation_function(df, output_dir, filename_suffix=''):
     """
     Given a time series DataFrames (constructed with `make_time_series`),
@@ -300,7 +300,12 @@ def plot_autocorrelation_function(df, output_dir, filename_suffix=''):
         # use statsmodels for partial autocorrelation
         from statsmodels.graphics.tsaplots import plot_pacf
         fig, ax = plt.subplots(figsize=(8,5))
-        plot_pacf(series, label=series.name, ax=ax, zero=False)
+
+        # for newer version of statsmodels, need to work out lags explicitly
+        lim = min(int(np.ceil(10 * np.log10(nobs))), nobs - 1)
+        lags = np.arange(1,lim)
+        nlags = lags.max(0)
+        plot_pacf(series, nlags=nlags, label=series.name, ax=ax, zero=False)
         plt.ylim([-1.0, 1.0])
         plt.xlabel('Lag')
         plt.ylabel('Partial Autocorrelation')
@@ -309,7 +314,7 @@ def plot_autocorrelation_function(df, output_dir, filename_suffix=''):
         output_filename = series.name + '-partial-autocorrelation-function' + filename_suffix + '.png'
         plt.savefig(os.path.join(output_dir, output_filename), dpi=DPI)
         plt.close(fig)
-        
+
     # make plots for selected columns
[…]
