### 2020-09-04 168c3fa
fix to check_timeout
