### 2020-09-10 44def03
adding rst documentation (#369)
