+++ pyveg/docs/source/index.rst
@@ -1,0 +1,21 @@
+
+Welcome to pyveg's documentation!
+=================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   ```{include} README.md
+    :relative-images:
+   ```
+   pyveg.scripts.rst
+   pyveg.src.rst
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
