+++ pyveg/docs/source/modules.rst
@@ -1,0 +1,7 @@
+pyveg
+=====
+
+.. toctree::
+   :maxdepth: 4
+
+   pyveg
