### 2020-09-10 6f27061
moving docs rst for readthedocs  (#369)
