### 2020-09-10 45b17ce
showing scripts and src files in the index.rst file  (#369)
