### 2020-09-10 5e2ae4f
moving rst files to docs/  (#369)
