### 2020-09-10 5fbadf5
simplify azure downloading
