### 2020-09-10 35debcb
quick check if markdown functions with rst  (#369)
