+++ pyveg/requirements.txt
@@ -12,6 +12,7 @@ pytest
 requests
 dateparser
 python-igraph
+pandas==1.1.1
 geopandas
 earthengine-api==0.1.222
 geetools
