### 2020-09-14 f7fac9d
add_get_time_diff_function
