### 2020-09-14 30bcee4
add historical data to summary stats
