+++ pyveg/src/analysis_preprocessing.py
@@ -17,6 +17,8 @@ from statsmodels.nonparametric.smoothers_lowess import lowess
 
 from pyveg.src.data_analysis_utils import write_to_json
 
+from pyveg.src.date_utils import get_time_diff
+
 try:
     from pyveg.src import azure_utils
 except:
@@ -233,7 +235,18 @@ def make_time_series(dfs):
     ts_df = ts_df.loc[:, ~ts_df.columns.str.contains("longitude_std", case=False)]
 
     assert ts_df.empty == False
-    return ts_df
+    ts_list = []
+    earliest_date = ts_df.iloc[0]["date"]
+    earliest_veg_date = ts_df[ts_df["S2_offset50_mean"].notna()].iloc[0]["date"]
+    if get_time_diff(earliest_veg_date,earliest_date) > 10:
+        ts_df_historic = ts_df[ts_df["date"] < earliest_veg_date][["date","mean_2m_air_temperature","total_precipitation"]]
+        ts_df = ts_df[ts_df["date"] >= earliest_veg_date]
+        ts_list.append(ts_df)
+        ts_list.append(ts_df_historic)
+    else :
+        ts_list.append(ts_df)
+
+    return ts_list
 
 
 def resample_time_series(series, period="MS"):
@@ -938,7 +951,12 @@ def preprocess_data(
     store_feature_vectors(dfs, output_dir)
 
     # average over sub-images
-    ts_df = make_time_series(dfs)
+    ts_list = make_time_series(dfs)
+    ts_df = ts_list[0]
+    if len(ts_list) > 1 :
+        ts_historic = ts_list[1]
+    else :
+        ts_historic = pd.DataFrame()
 
     # resample the averaged time series using linear interpolation
     if resample:
@@ -954,6 +972,10 @@ def preprocess_data(
     ts_filename = os.path.join(output_dir, "time_series.csv")
     print(f'- Saving time series to "{ts_filename}".')
     ts_df.to_csv(ts_filename, index=False)
+    if not ts_historic.empty :
+        ts_filename = os.path.join(output_dir, "time_series_historic.csv")
+        print(f'- Saving time series to "{ts_filename}".')
+        ts_historic.to_csv(ts_filename, index=False)
 
     # additionally save resampled & detrended time series
     # this detrending option (one year seasonality substraction) only works in monthly data that has at least 2 years of data
@@ -967,7 +989,7 @@ def preprocess_data(
         dfs_detrended_smooth = smooth_veg_data(dfs_detrended, n=12)
 
         # combine over sub-images
-        ts_df_detrended_smooth = make_time_series(dfs_detrended_smooth)
+        ts_df_detrended_smooth = make_time_series(dfs_detrended_smooth)[0]
 
         # save output
         ts_filename_detrended = os.path.join(output_dir, "time_series_detrended.csv")
@@ -1002,9 +1024,12 @@ def save_ts_summary_stats(ts_dirname, output_dir, metadata):
 
         # we should get one seasonal time series and a detrended one
         ts_df_detrended = pd.DataFrame()
+        ts_df_historic = pd.DataFrame()
         for filename in ts_filenames:
             if "detrended" in filename:
                 ts_df_detrended = pd.read_csv(os.path.join(ts_dirname,filename))
+            elif "historic" in filename:
+                ts_df_historic = pd.read_csv(os.path.join(ts_dirname,filename))
             else:
                 ts_df = pd.read_csv(os.path.join(ts_dirname,filename))
 
@@ -1021,13 +1046,22 @@ def save_ts_summary_stats(ts_dirname, output_dir, metadata):
 
             return stats_dict
 
+        # calculate summary statistics for each relevant time series
+        ts_dict_list = []
         # only look at relevant time series (offset50, ndvi and precipitation)
+        if not ts_df_historic.empty :
[…]
