### 2020-09-14 279df73
Scale NDVI and NDVI_Veg
