### 2020-09-15 df97566
updating index functions (#369)
