+++ docs/pyveg.scripts.rst
@@ -1,109 +1,0 @@
-pyveg.scripts package
-=====================
-
-Submodules
-----------
-
-pyveg.scripts.analyse\_gee\_data module
----------------------------------------
-
-.. automodule:: pyveg.scripts.analyse_gee_data
-   :members:
-   :undoc-members: pyveg.scripts.analyse_gee_data.main
-   :show-inheritance:
-
-pyveg.scripts.analyse\_pyveg\_summary\_data module
---------------------------------------------------
-
-.. automodule:: pyveg.scripts.analyse_pyveg_summary_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.calc\_euler\_characteristic module
-------------------------------------------------
-
-.. automodule:: pyveg.scripts.calc_euler_characteristic
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.create\_analysis\_report module
----------------------------------------------
-
-.. automodule:: pyveg.scripts.create_analysis_report
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.crop\_and\_convert\_images module
------------------------------------------------
-
-.. automodule:: pyveg.scripts.crop_and_convert_images
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.download\_from\_azure module
-------------------------------------------
-
-.. automodule:: pyveg.scripts.download_from_azure
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.generate\_config\_file module
--------------------------------------------
-
-.. automodule:: pyveg.scripts.generate_config_file
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.generate\_pattern module
---------------------------------------
-
-.. automodule:: pyveg.scripts.generate_pattern
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.plot\_feature\_vectors module
--------------------------------------------
-
-.. automodule:: pyveg.scripts.plot_feature_vectors
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.run\_pyveg\_module module
----------------------------------------
-
-.. automodule:: pyveg.scripts.run_pyveg_module
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.run\_pyveg\_pipeline module
------------------------------------------
-
-.. automodule:: pyveg.scripts.run_pyveg_pipeline
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyveg.scripts.upload\_to\_zenodo module
----------------------------------------
-
-.. automodule:: pyveg.scripts.upload_to_zenodo
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pyveg.scripts
-   :members:
-   :undoc-members:
-   :show-inheritance:
