+++ docs/source/index.rst
@@ -1,14 +1,12 @@
-
 Welcome to pyveg's documentation!
 =================================
+
+Test something
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
-   README.md
-   pyveg.scripts.rst
-   pyveg.src.rst
 
 
 Indices and tables
