+++ docs/pyveg.rst
@@ -1,20 +1,0 @@
-pyveg package
-=============
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   pyveg.scripts
-   pyveg.src
-
-
-Module contents
----------------
-
-.. automodule:: pyveg
-   :members:
-   :undoc-members:
-   :show-inheritance:
