+++ docs/pyveg.src.rst
@@ -1,149 +1,0 @@
-pyveg.src package
-=================
-
-Submodules
-----------
-
-pyveg.src.analysis\_preprocessing module
-----------------------------------------
-
-.. automodule:: pyveg.src.analysis_preprocessing
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.azure\_utils module
------------------------------
-
-.. automodule:: pyveg.src.azure_utils
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.batch\_utils module
------------------------------
-
-.. automodule:: pyveg.src.batch_utils
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.combiner\_modules module
-----------------------------------
-
-.. automodule:: pyveg.src.combiner_modules
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.coordinate\_utils module
-----------------------------------
-
-.. automodule:: pyveg.src.coordinate_utils
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.data\_analysis\_utils module
---------------------------------------
-
-.. automodule:: pyveg.src.data_analysis_utils
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.date\_utils module
-----------------------------
-
-.. automodule:: pyveg.src.date_utils
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.download\_modules module
-----------------------------------
-
-.. automodule:: pyveg.src.download_modules
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.file\_utils module
-----------------------------
-
-.. automodule:: pyveg.src.file_utils
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.gee\_interface module
--------------------------------
-
-.. automodule:: pyveg.src.gee_interface
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.image\_utils module
------------------------------
-
-.. automodule:: pyveg.src.image_utils
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.pattern\_generation module
-------------------------------------
-
-.. automodule:: pyveg.src.pattern_generation
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.plotting module
--------------------------
-
-.. automodule:: pyveg.src.plotting
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.processor\_modules module
------------------------------------
-
-.. automodule:: pyveg.src.processor_modules
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.pyveg\_pipeline module
---------------------------------
-
-.. automodule:: pyveg.src.pyveg_pipeline
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.subgraph\_centrality module
--------------------------------------
-
-.. automodule:: pyveg.src.subgraph_centrality
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-pyveg.src.zenodo\_utils module
-------------------------------
-
-.. automodule:: pyveg.src.zenodo_utils
-   :members:
-   :exclude-members: main
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pyveg.src
-   :members:
-   :exclude-members: main
-   :show-inheritance:
