+++ docs/modules.rst
@@ -1,7 +1,0 @@
-pyveg
-=====
-
-.. toctree::
-   :maxdepth: 4
-
-   pyveg
