### 2020-09-15 0cc8ef1
master_doc = 'index' in conf.py(#369)
