### 2020-09-15 a110be8
adding napoleon extention (#369)
