+++ docs/source/conf.py
@@ -30,7 +30,8 @@ master_doc = 'index'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = [
+extensions = [ 'sphinx.ext.napoleon',
+
 ]
 
 # Add any paths that contain templates here, relative to this directory.
