### 2020-09-15 28de484
Improve Scale NDVI and NDVI_Veg
