### 2020-09-15 b81d3ed
adding readthedocs yml file (#369)
