### 2020-09-15 f33cd1f
changing readthedocs yml file (#369)
