+++ .readthedocs.yml
@@ -19,7 +19,10 @@ formats:
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-  version: 3.7
-  install:
-    - method: setuptools
-    - path: .
+   version: 3.7
+   install:
+      - method: pip
+        path: .
+      - method: setuptools
+        path: .
+  system_packages: true
