### 2020-09-15 d3bd1ee
changing readthedocs yml file (#369)
