+++ .readthedocs.yml
@@ -23,6 +23,3 @@ python:
    install:
       - method: pip
         path: .
-      - method: setuptools
-        path: .
-  system_packages: true
