+++ .readthedocs.yml
@@ -23,3 +23,5 @@ python:
    install:
       - method: pip
         path: .
+
+requirements_file: docs/requirements.txt
