+++ docs/source/pyveg.src.rst
@@ -1,0 +1,149 @@
+pyveg.src package
+=================
+
+Submodules
+----------
+
+pyveg.src.analysis\_preprocessing module
+----------------------------------------
+
+.. automodule:: pyveg.src.analysis_preprocessing
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.azure\_utils module
+-----------------------------
+
+.. automodule:: pyveg.src.azure_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.batch\_utils module
+-----------------------------
+
+.. automodule:: pyveg.src.batch_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.combiner\_modules module
+----------------------------------
+
+.. automodule:: pyveg.src.combiner_modules
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.coordinate\_utils module
+----------------------------------
+
+.. automodule:: pyveg.src.coordinate_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.data\_analysis\_utils module
+--------------------------------------
+
+.. automodule:: pyveg.src.data_analysis_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.date\_utils module
+----------------------------
+
+.. automodule:: pyveg.src.date_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.download\_modules module
+----------------------------------
+
+.. automodule:: pyveg.src.download_modules
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.file\_utils module
+----------------------------
+
+.. automodule:: pyveg.src.file_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.gee\_interface module
+-------------------------------
+
+.. automodule:: pyveg.src.gee_interface
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.image\_utils module
+-----------------------------
+
+.. automodule:: pyveg.src.image_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.pattern\_generation module
+------------------------------------
+
+.. automodule:: pyveg.src.pattern_generation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.plotting module
+-------------------------
+
+.. automodule:: pyveg.src.plotting
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.processor\_modules module
+-----------------------------------
+
+.. automodule:: pyveg.src.processor_modules
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.pyveg\_pipeline module
+--------------------------------
+
+.. automodule:: pyveg.src.pyveg_pipeline
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.subgraph\_centrality module
+-------------------------------------
+
+.. automodule:: pyveg.src.subgraph_centrality
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyveg.src.zenodo\_utils module
+------------------------------
+
+.. automodule:: pyveg.src.zenodo_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyveg.src
+   :members:
+   :undoc-members:
+   :show-inheritance:
