### 2020-09-15 203de8d
removing requirements.txt docs (#369)
