+++ .readthedocs.yml
@@ -23,5 +23,3 @@ python:
    install:
       - method: pip
         path: .
-
-requirements_file: docs/requirements.txt
