+++ docs/source/index.rst
@@ -7,6 +7,7 @@ Test something
    :maxdepth: 2
    :caption: Contents:
 
+   README.md
    pyveg.scripts.rst
    pyveg.src.rst
 
