+++ docs/source/conf.py
@@ -46,11 +46,6 @@ extensions = [
 napoleon_use_ivar = True
 
 # for Sphinx-1.3
-from recommonmark.parser import CommonMarkParser
-
-source_parsers = {
-    '.md': CommonMarkParser,
-}
 
 source_suffix = ['.rst', '.md']
 # Add any paths that contain templates here, relative to this directory.
