### 2020-09-15 1140473
in make_time_series remove hardcoded column name
