### 2020-09-17 ada7ed6
remove code related to using the ImageLabeller output to mask sub-images
