### 2020-09-17 700cf22
beautifiying the readthedocs (#369)
