+++ docs/source/index.rst
@@ -1,13 +1,28 @@
 Welcome to pyveg's documentation!
 =================================
 
-Test something
+The `pyveg` package implements functionality to download and process data
+from Google Earth Engine (GEE), and analyse images from periodic vegetation patterns (PVP) in arid and semi-arid ecosystems.
+PVP images are quantified using
+network centrality metrics. The results of the analysis can be used
+to search for typical early warning signals of an ecological collapse. Google Earth Engine Editor scripts are also provided to help
+researchers discover locations of ecosystems which may be in
+decline.
+
+pyveg is being developed as part of a research project looking for evidence of early warning signals of ecosystem
+collapse using remote sensing data. pyveg allows such research to be carried out at scale, and hence can be an important
+tool in understanding changing arid and semi-arid ecosystem dynamics.
 
 .. toctree::
    :maxdepth: 2
-   :caption: Contents:
+   :caption: The pyveg package
 
    README.md
+
+.. toctree::
+   :maxdepth: 2
+   :caption: API documentation
+
    pyveg.scripts.rst
    pyveg.src.rst
 
