### 2020-09-18 645bc4b
Update analyse_pyveg_summary_data.py
