### 2020-09-18 dfbcb5c
Update analyse_pyveg_summary_data.py
