+++ README.md
@@ -2,6 +2,8 @@
 
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/alan-turing-institute/monitoring-ecosystem-resilience/develop?filepath=notebooks)
 
+[![Documentation Status](https://readthedocs.org/projects/pyveg/badge/?version=latest)](https://pyveg.readthedocs.io/en/latest/?badge=latest)
+
 # monitoring-ecosystem-resilience
 Repository for mini-projects in the Data science for Sustainable development project.
 
