+++ pyveg/coordinates.py
@@ -54,14 +54,14 @@ coordinate_store.loc["02"] = {
 coordinate_store.loc["03"] = {
     "continent": "Africa",
     "country": "Niger",
-    "type": "tiger bush",
+    "type": "labyrinths",
     "latitude": 13.12,
     "longitude": 2.59,
 }
 coordinate_store.loc["04"] = {
     "continent": "Africa",
     "country": "Niger",
-    "type": "tiger bush",
+    "type": "labyrinths",
     "latitude": 13.17,
     "longitude": 1.58,
 }
@@ -136,7 +136,7 @@ coordinate_store.loc["12"] = {
 coordinate_store.loc["13"] = {
     "continent": "America",
     "country": "Mexico",
-    "type": "tiger bush",
+    "type": "labyrinths",
     "latitude": 26.82,
     "longitude": -112.86,
 }
@@ -154,14 +154,14 @@ coordinate_store.loc["14"] = {
 coordinate_store.loc["15"] = {
     "continent": "Africa",
     "country": "Niger",
-    "type": "declining",
+    "type": "labyrinths",
     "latitude": 14.91,
     "longitude": -0.66,
 }
 coordinate_store.loc["16"] = {
     "continent": "Africa",
     "country": "Niger",
-    "type": "declining",
+    "type": "spots",
     "latitude": 15.03,
     "longitude": -0.87,
 }
@@ -175,42 +175,42 @@ coordinate_store.loc["17"] = {
 coordinate_store.loc["18"] = {
     "continent": "Africa",
     "country": "Niger",
-    "type": "declining",
+    "type": "spots/labyrinths",
     "latitude": 15.34,
     "longitude": -1.15,
 }
 coordinate_store.loc["19"] = {
     "continent": "Africa",
-    "country": "Niger",
+    "country": "Mali",
     "type": "declining",
     "latitude": 15.14,
     "longitude": -1.16,
 }
 coordinate_store.loc["20"] = {
     "continent": "Africa",
-    "country": "Niger",
-    "type": "declining",
+    "country": "Mali",
+    "type": "spots/labyrinths",
     "latitude": 14.85,
     "longitude": -1.43,
 }
 coordinate_store.loc["21"] = {
     "continent": "Africa",
-    "country": "Niger",
-    "type": "declining",
+    "country": "Mali",
+    "type": "spots/labyrinths",
     "latitude": 14.97,
     "longitude": -1.12,
 }
 coordinate_store.loc["22"] = {
     "continent": "Africa",
-    "country": "Niger",
-    "type": "declining",
+    "country": "Mali",
+    "type": "labyrinths",
     "latitude": 15.14,
     "longitude": -1.56,
 }
 coordinate_store.loc["23"] = {
     "continent": "Africa",
     "country": "Niger",
-    "type": "declining",
+    "type": "spots/labyrinths",
     "latitude": 15.02,
     "longitude": -1.35,
 }
@@ -218,43 +218,43 @@ coordinate_store.loc["23"] = {
 # degraded locations in West Niger
 coordinate_store.loc["24"] = {
     "continent": "Africa",
-    "country": "Niger",
+    "country": "Mali",
     "type": "degraded",
     "latitude": 16.26,
     "longitude": -1.83,
 }
 coordinate_store.loc["25"] = {
     "continent": "Africa",
-    "country": "Niger",
-    "type": "degraded",
+    "country": "Mali",
[…]
