### 2020-09-21 64e88fa
updating the README config explanation [WIP] (#369)
