### 2020-09-21 01116dd
updating the README figure (#369)
