+++ pyveg/README.md
@@ -11,7 +11,7 @@ conda activate veg
 ```
 Get the source.
 ```
-git clone https://alan-turing-institute/monitoring-ecosystem-resilience.git
+git clone https://github.com/alan-turing-institute/monitoring-ecosystem-resilience.git
 ```
 Enter the repository and check out a relevant branch if necessary (the `develop` branch contains the most
 up to date stable version of the code).
