### 2020-09-22 06f0162
add function to get the currently checked out git tag
