### 2020-09-22 d332402
shorten long filename of summary_stats csv
