### 2020-09-22 92a7cec
allow analysis to be run off zenodo summary json
