### 2020-09-22 eefc257
streamline upload_to_zenodo to only deal with results_summary.json
