+++ pyveg/scripts/upload_to_zenodo.py
@@ -1,5 +1,5 @@
 """
-Upload the results_summary.json, and the outputs of the time-series analysis
+Upload the results_summary.json or the outputs of the time-series analysis
 to the Zenodo open source platform for data www.zenodo.org.
 
 Will create a zipfile, with a name based upon the coordinates and
@@ -11,7 +11,11 @@ for testing.  Use the '--test_api' argument to use this.
 
 import os
 import argparse
+import tempfile
+import json
 
+from pyveg.src.file_utils import get_tag, construct_filename_from_metadata
+from pyveg.src.analysis_preprocessing import read_results_summary
 from pyveg.src.zenodo_utils import (
     get_deposition_id,
     prepare_results_zipfile,
@@ -42,30 +46,32 @@ def create_new_deposition(use_sandbox=False):
     return deposition_id
 
 
-def upload_results(collection,
-                   png_location,
-                   png_location_type,
-                   json_location,
-                   json_location_type,
-                   use_test_api):
+def upload_results_summary(json_location,
+                           json_location_type,
+                           use_test_api):
     deposition_id = get_deposition_id(use_test_api)
-    zipfile = prepare_results_zipfile(collection,
-                                      png_location,
-                                      png_location_type,
-                                      json_location,
-                                      json_location_type )
-    uploaded_ok = upload_file(zipfile, deposition_id, use_test_api)
+
+    # read in the json
+    results_summary = read_results_summary(json_location, input_location_type=json_location_type)
+    if (not results_summary) or (not "metadata" in results_summary.keys()):
+        print("Unable to find metadata in json file in {}".format(json_location))
+        return False
+    filename = construct_filename_from_metadata(results_summary["metadata"],"results_summary.json")
+    tmpdir = tempfile.mkdtemp()
+    filepath = os.path.join(tmpdir, filename)
+    with open(filepath, "w") as outfile:
+        json.dump(results_summary, outfile)
+    print("Uploading {} to Zenodo".format(filename))
+    uploaded_ok = upload_file(filepath, deposition_id, use_test_api)
+
     return uploaded_ok
 
 
 def main():
     parser = argparse.ArgumentParser(description="upload to zenodo")
     parser.add_argument("--create_deposition", help="create a new deposition", action="store_true")
-    parser.add_argument("--input_png_loc",help="path to analysis/ subdirectory")
-    parser.add_argument("--png_loc_type",help="'local' or 'azure'", default="local")
-    parser.add_argument("--input_json_loc",help="path to dir containing results_summary.json")
-    parser.add_argument("--json_loc_type",help="'local' or 'azure'", default="local")
-    parser.add_argument("--collection", help="Collection name, e.g. 'Sentinel2'")
+    parser.add_argument("--input_location",help="directory or container with file of interest",required=True)
+    parser.add_argument("--input_location_type",help="'local' or 'azure'", default="local")
     parser.add_argument("--test_api", help="use the test API", action="store_true")
     args = parser.parse_args()
     if args.create_deposition:
@@ -82,23 +88,11 @@ We normally do this just once per project.  Please confirm (y/n)
         else:
             print("Confirmation not received - exiting with no action")
             return
-    if not args.input_png_loc and args.collection:
-        raise RuntimeError("--input_png_loc and --collection are required arguments for uploading data")
-    png_location = args.input_png_loc
-    png_location_type = args.png_loc_type
-    if args.input_json_loc:
-        json_location = args.input_json_loc
-    else:
-        json_location = png_location
-    json_loc_type = args.json_loc_type
-    collection = args.collection,
+
     test_api = True if args.test_api else False
-    result = upload_results(collection,
-                            png_location,
-                            png_location_type,
-                            json_location,
-                            json_location_type,
[…]
