### 2020-09-22 3a20718
add function to generate filenames
