### 2020-09-22 b44c065
streamline the long filename for summary stats csv
