### 2020-09-23 f450ed6
enable coords_id to go into config file
