### 2020-09-23 f0fdb33
put tag and coords_id into metadata
