### 2020-09-23 23d86a0
add region size to config file
