+++ pyveg/coordinates.py
@@ -20,7 +20,7 @@ import pandas as pd
 
 # initialise a DataFrame to store coordinate
 coordinate_store = pd.DataFrame(
-    columns=["continent", "country", "type", "latitude", "longitude"]
+    columns=["continent", "country", "type", "latitude", "longitude", "region_size"]
 )
 
 # rename index
@@ -34,6 +34,7 @@ coordinate_store.loc["00"] = {
     "type": "spots",
     "latitude": 11.58,
     "longitude": 27.94,
+    "region_size": 0.08
 }
 coordinate_store.loc["01"] = {
     "continent": "Africa",
@@ -41,6 +42,7 @@ coordinate_store.loc["01"] = {
     "type": "labyrinths",
     "latitude": 11.12,
     "longitude": 28.37,
+    "region_size": 0.08
 }
 coordinate_store.loc["02"] = {
     "continent": "Africa",
@@ -48,6 +50,7 @@ coordinate_store.loc["02"] = {
     "type": "gaps",
     "latitude": 10.96,
     "longitude": 28.20,
+    "region_size": 0.08
 }
 
 # niger
@@ -57,6 +60,7 @@ coordinate_store.loc["03"] = {
     "type": "tiger bush",
     "latitude": 13.12,
     "longitude": 2.59,
+    "region_size": 0.08
 }
 coordinate_store.loc["04"] = {
     "continent": "Africa",
@@ -64,6 +68,7 @@ coordinate_store.loc["04"] = {
     "type": "tiger bush",
     "latitude": 13.17,
     "longitude": 1.58,
+    "region_size": 0.08
 }
 
 # senegal
@@ -73,6 +78,7 @@ coordinate_store.loc["05"] = {
     "type": "labyrinths",
     "latitude": 15.20,
     "longitude": -15.20,
+    "region_size": 0.08
 }
 coordinate_store.loc["06"] = {
     "continent": "Africa",
@@ -80,6 +86,7 @@ coordinate_store.loc["06"] = {
     "type": "labyrinths",
     "latitude": 15.09,
     "longitude": -15.04,
+    "region_size": 0.08
 }
 coordinate_store.loc["07"] = {
     "continent": "Africa",
@@ -87,6 +94,7 @@ coordinate_store.loc["07"] = {
     "type": "gaps",
     "latitude": 15.80,
     "longitude": -14.36,
+    "region_size": 0.08
 }
 coordinate_store.loc["08"] = {
     "continent": "Africa",
@@ -94,6 +102,7 @@ coordinate_store.loc["08"] = {
     "type": "gaps",
     "latitude": 15.11,
     "longitude": -14.53,
+    "region_size": 0.08
 }
 
 # zambia
@@ -103,6 +112,7 @@ coordinate_store.loc["09"] = {
     "type": "gaps",
     "latitude": -15.34,
     "longitude": 22.22,
+    "region_size": 0.08
 }
 
 # kenya
@@ -112,6 +122,7 @@ coordinate_store.loc["10"] = {
     "type": "spots",
     "latitude": 0.43,
     "longitude": 40.30,
+    "region_size": 0.08
 }
 
 # somalia
@@ -121,6 +132,7 @@ coordinate_store.loc["11"] = {
     "type": "labyrinths",
     "latitude": 8.09,
     "longitude": 47.44,
+    "region_size": 0.08
 }
 
 # australia
@@ -130,6 +142,7 @@ coordinate_store.loc["12"] = {
     "type": "gaps",
     "latitude": -15.71,
     "longitude": 133.10,
+    "region_size": 0.08
 }
 
 # usa
@@ -139,6 +152,7 @@ coordinate_store.loc["13"] = {
[…]
