### 2020-09-24 0ea301a
adding contributing, Using Azure and Zenodo guidelines (#369)
