+++ pyveg/README.md
@@ -14,6 +14,8 @@ The code in this repository is intended to perform two main tasks:
 the pattern vegetation are then calculated on the sub-image level. Both colour (RGB) and Normalised Difference Vegetation 
 Index (NDVI) images are downloaded and stored on the sub-image level. 
     * For weather collections the precipitation and temperature "images" are averaged into a single value at each point in the time series.
+* The download job is fully specified by a configuration file that can be generated by specifying the details of the 
+data they want to download via prompts (satelite to use, coordinates, time period, number of time points, etc.).  
 
 **Time series analysis on downloaded data**:
 
@@ -35,8 +37,11 @@ The digram below represents the high level flow of the `pyveg` package.
 
 ![`pyveg` program flow.](paper/pveg_flow.png)
 
+The full ReadTheDocs documentation for this `pyveg` can be found in this [link](https://pyveg.readthedocs.io/en/latest/).
+
 
 This page contains an installation guide, and some usage examples for this package.
+
 
 ## Installation
 
@@ -94,7 +99,7 @@ to identify the download job using the `--name` argument.
 
 Note that we use the GEE convention for coordinates, i.e. `(longitude,latitude)`.
 
-#### Generating a donwload configuration file 
+#### Generating a download configuration file 
 
 To create a configuration file for use in the pyveg pipeline described above, use the command 
 ```
