### 2020-09-24 c2696ea
adding more work on the Readme and contributing guidelines (#369)
