### 2020-09-24 aae9666
new pyveg diagram (#369)
