### 2020-09-24 966df38
more text refactoring (#369)
