+++ pyveg/README.md
@@ -7,7 +7,7 @@ Google Earth Engine.
 
 The code in this repository is intended to perform two main tasks:
 
-**Download an process GEE data**:
+**1. Download an process GEE data**:
 
 * Download satellite data from Google Earth Engine (images and weather data).
     * Downloaded images are divided into 50x50 pixel sub-images, network centrality metrics are used to describe
@@ -17,7 +17,7 @@ Index (NDVI) images are downloaded and stored on the sub-image level.
 * The download job is fully specified by a configuration file that can be generated by specifying the details of the 
 data they want to download via prompts (satelite to use, coordinates, time period, number of time points, etc.).  
 
-**Time series analysis on downloaded data**:
+**2. Time series analysis on downloaded data**:
 
 * Time series analysis of following metrics: raw NDVI mean pixel intensity across the image, vegetation network centrality metric, and precipitation.
     * The time series are processed  (outliers removed and resampled to avoid gaps). All time series of each
@@ -54,7 +54,7 @@ The full ReadTheDocs documentation for this `pyveg` can be found in this [link](
 This page contains an installation guide, and some usage examples for this package.
 
 
-## Installation
+# Installation
 
 `pyveg` requires Python 3.6 or greater. To install, start by creating a fresh `conda` environment.
 ```
