### 2020-09-24 f949f19
spelling checks (#369)
