+++ pyveg/README.md
@@ -15,17 +15,17 @@ the pattern vegetation are then calculated on the sub-image level. Both colour (
 Index (NDVI) images are downloaded and stored on the sub-image level. 
     * For weather collections the precipitation and temperature "images" are averaged into a single value at each point in the time series.
 * The download job is fully specified by a configuration file that can be generated by specifying the details of the 
-data they want to download via prompts (satelite to use, coordinates, time period, number of time points, etc.).  
+data they want to download via prompts (satellite to use, coordinates, time period, number of time points, etc.).  
 
 **2. Time series analysis on downloaded data**:
 
 * Time series analysis of following metrics: raw NDVI mean pixel intensity across the image, vegetation network centrality metric, and precipitation.
     * The time series are processed  (outliers removed and resampled to avoid gaps). All time series of each
      sub-image are aggregated into one summary time series that is used for analysis. 
-    * The summary time series is deasonalised and smoothed time series and residuals between the raw and de-seasonalised
+    * The summary time series is de-seasonalised and smoothed time series and residuals between the raw and de-seasonalised
      and smoothed time series are calculated and used for an early warning resilience analysis.
 
-* Time series plots are produced, along with auto- and cross-correlation plots. Early warning signals are also computed using the ewstools package [@ewstools], 
+* Time series plots are produced, along with auto- and cross-correlation plots. Early warning signals are also computed using the [ewstools package](https://github.com/ThomasMBury/ewstools), 
 including Lag-1 autocorrelation and standard deviation moving window plots. A sensitivity and significance analysis is 
 also performed in order to determine whether any trends are statistically significant.
 * Time series summary statistics and resilience metrics are saved into files.
@@ -43,7 +43,7 @@ pipeline (downloading + time series analysis).
 ### `pyveg` flow
 
 The digram below represents the high level flow of the main functionalities of the
-`pyveg` package. For each main component there is an CLI console_scripts defined, that is shown
+`pyveg` package. For each main component there is an CLI console scripts defined, that is shown
 in the diagram. 
 
 ![The`pyveg` program flow.](paper/pveg_flow.png)
@@ -54,7 +54,7 @@ The full ReadTheDocs documentation for this `pyveg` can be found in this [link](
 This page contains an installation guide, and some usage examples for this package.
 
 
-# Installation
+## Installation
 
 `pyveg` requires Python 3.6 or greater. To install, start by creating a fresh `conda` environment.
 ```
@@ -86,7 +86,7 @@ A new browser window will open. Copy the token from this window to the terminal 
 complete the authentication process.
 
 
-## Google Earth Engine
+### Google Earth Engine
 
 [Google Earth Engine](https://earthengine.google.com) (GEE) is a powerful tool for obtaining 
 and analysing satellite imagery. This directory contains some useful scripts for interacting 
@@ -197,7 +197,8 @@ If you have access to Microsoft Azure cloud computing facilities, downloading an
 ### Downloading data using the API
 
 Although `pyveg` has been mostly designed to be used with the CLI as shown above, we can also use `pyveg` functions through 
-the API. A tutorial of how to download data this way is included in [here](notebooks/tutorial_download_and_process_gee_images.ipynb). 
+the API. A tutorial of how to download data this way is included in the ``notebooks/tutorial_download_and_process_gee_images.ipynb`` notebook
+tutorial. 
 
 ## Analysing the Downloaded Data with `pyveg`
 
@@ -225,36 +226,37 @@ The analysis script ran with the ```pyveg_gee_analysis``` commands run the follo
 
 #### Preprocessing
 
-`pyevg` supports the following preprocessing operations:
[…]
