### 2020-09-24 13b5b42
refactoring text (#369)
