### 2020-09-24 bd92853
refactoring text (#369)
