### 2020-09-24 59ff2e4
adding new authors for pyveg (#369)
