### 2020-09-24 ec6a431
a few changes to the readme text
