+++ pyveg/README.md
@@ -2,50 +2,40 @@
 
 ## Introduction 
 
-The `pyveg` package is developed to study the evolution of vegetation patterns in semi-arid environments using
-Google Earth Engine.
+The `pyveg` package is developed to study the evolution of vegetation patterns in semi-arid environments using data downloaded from Google Earth Engine.
 
 The code in this repository is intended to perform two main tasks:
 
-**1. Download an process GEE data**:
+**1. Download and process GEE data**:
 
 * Download satellite data from Google Earth Engine (images and weather data).
-    * Downloaded images are divided into 50x50 pixel sub-images, network centrality metrics are used to describe
-the pattern vegetation are then calculated on the sub-image level. Both colour (RGB) and Normalised Difference Vegetation 
-Index (NDVI) images are downloaded and stored on the sub-image level. 
+    * Downloaded images are divided into 50x50 pixel sub-images, network centrality metrics are used to describe the pattern vegetation are then calculated on the sub-image level. Both colour (RGB) and Normalised Difference Vegetation Index (NDVI) images are downloaded and stored on the sub-image level. 
     * For weather collections the precipitation and temperature "images" are averaged into a single value at each point in the time series.
-* The download job is fully specified by a configuration file that can be generated by specifying the details of the 
-data they want to download via prompts (satellite to use, coordinates, time period, number of time points, etc.).  
+* The download job is fully specified by a configuration file that can be generated by specifying the details of the data to be downloaded via prompts (satellite to use, coordinates, time period, number of time points, etc.).  
 
 **2. Time series analysis on downloaded data**:
 
-* Time series analysis of following metrics: raw NDVI mean pixel intensity across the image, vegetation network centrality metric, and precipitation.
-    * The time series are processed  (outliers removed and resampled to avoid gaps). All time series of each
-     sub-image are aggregated into one summary time series that is used for analysis. 
-    * The summary time series is de-seasonalised and smoothed time series and residuals between the raw and de-seasonalised
-     and smoothed time series are calculated and used for an early warning resilience analysis.
+* Time series analysis of the following metrics: raw NDVI mean pixel intensity across the image, vegetation network centrality metric, and precipitation.
+    * The time series are processed (outliers removed and resampled to avoid gaps). All time series of each sub-image are aggregated into one summary time series that is used for analysis. 
+    * The summary time series is de-seasonalised and smoothed.
+    * Residuals between the raw and de-seasonalised and smoothed time series are calculated and used for an early warning resilience analysis.
 
-* Time series plots are produced, along with auto- and cross-correlation plots. Early warning signals are also computed using the [ewstools package](https://github.com/ThomasMBury/ewstools), 
-including Lag-1 autocorrelation and standard deviation moving window plots. A sensitivity and significance analysis is 
-also performed in order to determine whether any trends are statistically significant.
+* Time series plots are produced, along with auto- and cross-correlation plots. Early warning signals are also computed using the [ewstools package](https://github.com/ThomasMBury/ewstools), including Lag-1 autocorrelation and standard deviation moving window plots. A sensitivity and significance analysis is also performed in order to determine whether any trends are statistically significant.
 * Time series summary statistics and resilience metrics are saved into files.
-* A PDF report is created with showcasing the main figures resulting of this analysis. 
+* A PDF report is created showcasing the main figures resulting from the analyses. 
 
 **Other functionalities**:
 
 `pyveg` also has other minor functionalities:
 
[…]
