### 2020-09-25 f802f2c
add pattern type to metadata
