+++ pyveg/scripts/generate_config_file.py
@@ -79,6 +79,7 @@ def make_filename(configs_dir,
                   longitude,
                   latitude,
                   country,
+                  pattern_type,
                   start_date,
                   end_date,
                   time_per_point,
@@ -94,7 +95,7 @@ def make_filename(configs_dir,
         filename_start += "_"+coords_id
     filepath = os.path.join(
         configs_dir,
-        f"{filename_start}_{collection_name}_{latitude}N_{longitude}E_{country}_{region_size}_{start_date}_{end_date}_{time_per_point}_{run_mode}.py"
+        f"{filename_start}_{collection_name}_{latitude}N_{longitude}E_{country}_{region_size}_{pattern_type}_{start_date}_{end_date}_{time_per_point}_{run_mode}.py"
         )
     return filepath
 
@@ -104,6 +105,7 @@ def write_file(configs_dir,
                longitude,
                latitude,
                country,
+               pattern_type,
                start_date,
                end_date,
                time_per_point,
@@ -122,6 +124,7 @@ def write_file(configs_dir,
                              longitude,
                              latitude,
                              country,
+                             pattern_type,
                              start_date,
                              end_date,
                              time_per_point,
@@ -143,27 +146,28 @@ def write_file(configs_dir,
 
     text = get_template_text()
     current_time = time.strftime("%y-%m-%d %H:%M:%S")
-    text = re.sub("CURRENT_TIME", current_time, text)
+    text = text.replace("CURRENT_TIME", current_time)
     output_location_type = "azure" if run_mode == "batch" else "local"
-    text = re.sub("COLLECTION_NAME", collection_name, text)
-    text = re.sub("WEATHER_COLL_NAME", weather_collection_name, text)
-    text = re.sub("OUTPUT_LOCATION_TYPE", output_location_type, text)
+    text = text.replace("COLLECTION_NAME", collection_name)
+    text = text.replace("WEATHER_COLL_NAME", weather_collection_name)
+    text = text.replace("OUTPUT_LOCATION_TYPE", output_location_type)
     text = text.replace("OUTPUT_LOCATION",output_location)
-    text = re.sub("LATITUDE", latitude, text)
-    text = re.sub("LONGITUDE", longitude, text)
-    text = re.sub("START_DATE", start_date, text)
-    text = re.sub("WEATHER_STARTDATE", weather_start_date, text)
-    text = re.sub("END_DATE", end_date, text)
-    text = re.sub("TIME_PER_POINT", time_per_point, text)
-    text = re.sub("REGION_SIZE", region_size, text)
-    text = re.sub("RUN_MODE", run_mode, text)
-    text = re.sub("NUM_THREADS", str(n_threads), text)
+    text = text.replace("LATITUDE", latitude)
+    text = text.replace("LONGITUDE", longitude)
+    text = text.replace("PATTERN_TYPE", pattern_type)
+    text = text.replace("START_DATE", start_date)
+    text = text.replace("WEATHER_STARTDATE", weather_start_date)
+    text = text.replace("END_DATE", end_date)
+    text = text.replace("TIME_PER_POINT", time_per_point)
+    text = text.replace("REGION_SIZE", region_size)
+    text = text.replace("RUN_MODE", run_mode)
+    text = text.replace("NUM_THREADS", str(n_threads))
     n_subimages = '10' if test_mode else '-1'
-    text = re.sub("NUM_SUBIMAGES", n_subimages, text)
+    text = text.replace("NUM_SUBIMAGES", n_subimages)
     if coords_id:
-        text = re.sub("COORDS_ID_STRING", 'coords_id = "{}"'.format(coords_id), text)
[…]
