### 2020-09-25 f02a42f
remove spaces from pattern_type strings
