### 2020-09-28 #2483 @crangelsmith
hi @labarba @arbennett, sorry for the silence. We have been working hard on the project, implementing the review comments but also in general development in view for a science paper we are also aiming to publish.

We are planning to do a PR to master this week, which will include the implementation of the comments from @arbennett. But also, the code has evolved a bit, and we would like to add a couple of more authors to the paper, to acknowledge the work some new members have done in the last couple of months. Is it ok to update the paper with this new info (authors + a couple of sentences about the new functionalities)?
