### 2020-09-28 alan-turing-institute/monitoring-ecosystem-resilience#371 @crangelsmith
This is implemented in the new version of the code. Will close now.
