### 2020-09-28 alan-turing-institute/monitoring-ecosystem-resilience#369 @crangelsmith
This is now implemented in the new READE.md and https://pyveg.readthedocs.io/en/latest/. Will close now.
