### 2020-09-28 alan-turing-institute/monitoring-ecosystem-resilience#370 @crangelsmith
This is implemented in the new README.md and https://pyveg.readthedocs.io/en/latest/. Will close the issue now.
