### 2020-09-28 52564bb
updating tutorial_analyse_gee_data.ipynb with new options for config generation (#367)
