### 2020-09-28 f4ef9ba
make a warning message a bit more informative
