### 2020-09-29 ea90d70
add function to delete files from zenodo
