### 2020-10-01 6a850fb
implementing the cb fit in the summary stats (#455)
