### 2020-10-01 098b43e
Add extra plots and functions
