+++ pyveg/scripts/analyse_pyveg_summary_data.py
@@ -19,103 +19,129 @@ from pyveg.scripts.upload_to_zenodo import upload_results
 def barplot_plots(df, output_dir):
 
     plt.figure()
-    ax13 = sns.barplot(x='name',y='S2_offset50_mean_max',hue='total_precipitation_mean',data=df)
-    ax13.set_xlabel("Mean precipitation over time series")
-    ax13.set_ylabel("Max Offset50 over time series")
+    ax20 = sns.barplot(x='name',y='S2_offset50_mean_max',hue='total_precipitation_mean',data=df)
+    ax20.set_xlabel("Mean precipitation over time series")
+    ax20.set_ylabel("Max Offset50 over time series")
     plt.savefig(os.path.join(output_dir,'offset50_precipitation_bar.png'))
 
     plt.figure()
-    ax14 = sns.barplot(x='name',y='S2_offset50_mean_max',hue='S2_offset50_mean_Lag-1 AC (0.99 rolling window)',data=df)
-    ax14.set_xlabel("Offset50 Lag-1 AC (0.99 rolling window)")
-    ax14.set_ylabel("Max Offset50 over time series")
+    ax20 = sns.barplot(x='name',y='S2_offset50_mean_max',hue='S2_offset50_mean_Lag-1 AC (0.99 rolling window)',data=df)
+    ax20.set_xlabel("Offset50 Lag-1 AC (0.99 rolling window)")
+    ax20.set_ylabel("Max Offset50 over time series")
     plt.savefig(os.path.join(output_dir,'offset50_lag1ACvalue_bar.png'))
 
 
 def scatter_plots(df, output_dir):
 
     plt.figure()
-    ax = sns.scatterplot(y="S2_offset50_mean_mean", x="total_precipitation_mean", data=df)
+    ax = sns.scatterplot(y="S2_offset50_mean_mean", x="total_precipitation_mean", data=df,hue="S2_offset50_mean_pattern_type",palette="Accent_r",edgecolor="k",linewidth=1)
     ax.set_xlabel("Mean precipitation over time series")
     ax.set_ylabel("Mean Offset50 over time series")
     plt.savefig(os.path.join(output_dir,'offset50_vs_precipitation.png'))
 
     plt.figure()
-    ax1 = sns.scatterplot(x="longitude", y="S2_offset50_mean_mean", data=df)
+    ax1 = sns.scatterplot(x="longitude", y="S2_offset50_mean_mean", data=df,hue="S2_offset50_mean_pattern_type",palette="Accent_r",edgecolor="k",linewidth=1)
     ax1.set_ylabel("Mean Offset50 over time series")
     ax1.set_xlabel("Longitude")
     plt.savefig(os.path.join(output_dir,'offset50_vs_Longitude.png'))
 
     plt.figure()
-    ax2 = sns.scatterplot(x="latitude", y="S2_offset50_mean_mean", data=df)
+    ax2 = sns.scatterplot(x="latitude", y="S2_offset50_mean_mean", data=df,hue="S2_offset50_mean_pattern_type",palette="Accent_r",edgecolor="k",linewidth=1)
     ax2.set_ylabel("Mean Offset50 over time series")
     ax2.set_xlabel("Latitude")
     plt.savefig(os.path.join(output_dir,'offset50_vs_Latitude.png'))
 
     plt.figure()
-    ax3 = sns.scatterplot(x="longitude", y="latitude", size="S2_offset50_mean_mean", data=df)
+    ax3 = sns.scatterplot(x="longitude", y="latitude", size="S2_offset50_mean_mean", data=df,hue="S2_offset50_mean_pattern_type",palette="Accent_r",edgecolor="k",linewidth=1)
     ax3.set_ylabel("Latitude")
[…]
