### 2020-10-02 ea7a0e8
adding AR1 variance estimations for seasonal time series
