### 2020-10-02 alan-turing-institute/monitoring-ecosystem-resilience#368 @crangelsmith
This issue has been fixed and merged into master.
