### 2020-10-02 7fe2204
Calculate return rate and add plots
