+++ pyveg/scripts/analyse_pyveg_summary_data.py
@@ -15,7 +15,7 @@ import seaborn as sns
 import matplotlib.pyplot as plt
 from scipy.stats import pearsonr
 sns.set_style("whitegrid")
-from pyveg.scripts.upload_to_zenodo import upload_results
+#from pyveg.scripts.upload_to_zenodo import upload_results
 
 def barplot_plots(df, output_dir):
 
@@ -31,16 +31,17 @@ def barplot_plots(df, output_dir):
     """
     
     plt.figure()
-    ax20 = sns.barplot(x='name',y='S2_offset50_mean_max',hue='total_precipitation_mean',data=df)
-    ax20.set_xlabel("Mean precipitation over time series")
-    ax20.set_ylabel("Max Offset50 over time series")
+    ax30 = sns.barplot(x='name',y='S2_offset50_mean_max',hue='total_precipitation_mean',data=df)
+    ax30.set_xlabel("Mean precipitation over time series")
+    ax30.set_ylabel("Max Offset50 over time series")
     plt.savefig(os.path.join(output_dir,'offset50_precipitation_bar.png'))
 
     plt.figure()
-    ax20 = sns.barplot(x='name',y='S2_offset50_mean_max',hue='S2_offset50_mean_Lag-1 AC (0.99 rolling window)',data=df)
-    ax20.set_xlabel("Offset50 Lag-1 AC (0.99 rolling window)")
-    ax20.set_ylabel("Max Offset50 over time series")
+    ax31 = sns.barplot(x='name',y='S2_offset50_mean_max',hue='S2_offset50_mean_Lag-1 AC (0.99 rolling window)',data=df)
+    ax31.set_xlabel("Offset50 Lag-1 AC (0.99 rolling window)")
+    ax31.set_ylabel("Max Offset50 over time series")
     plt.savefig(os.path.join(output_dir,'offset50_lag1ACvalue_bar.png'))
+
 
 
 def scatter_plots(df, output_dir):
@@ -55,7 +56,7 @@ def scatter_plots(df, output_dir):
     output_dir : str
         Path to the directory to save plots to.
     """
-    
+
     plt.figure()
     ax = sns.scatterplot(y="S2_offset50_mean_mean", x="total_precipitation_mean", data=df,hue="S2_offset50_mean_pattern_type",palette="Accent_r",edgecolor="k",linewidth=1)
     ax.set_xlabel("Mean precipitation over time series")
@@ -138,13 +139,70 @@ def scatter_plots(df, output_dir):
     ax13 = sns.scatterplot(y="S2_offset50_mean_max", x="S2_ndvi_mean_max", data=df, hue="S2_offset50_mean_pattern_type",palette="Accent_r",edgecolor="k",linewidth=1)
     ax13.set_ylabel("Max offset50 over time series")
     ax13.set_xlabel("Max NDVI over time series")
-    plt.savefig(os.path.join(output_dir, 'max_offset50_vs_max_NDVI.png'))
+    plt.savefig(os.path.join(output_dir, 'offset50max_vs_NDVImax.png'))
 
     plt.figure()
     ax14 = sns.scatterplot(y="S2_offset50_mean_mean", x="S2_ndvi_mean_mean", data=df, hue="S2_offset50_mean_pattern_type",palette="Accent_r",edgecolor="k",linewidth=1)
     ax14.set_ylabel("Mean offset50 over time series")
     ax14.set_xlabel("Mean NDVI over time series")
-    plt.savefig(os.path.join(output_dir, 'mean_offset50_vs_mean_NDVI.png'))
+    plt.savefig(os.path.join(output_dir, 'offset50max_vs_NDVImean.png'))
+
+    plt.figure()
[…]
