### 2020-10-02 d208144
allow pyveg_analysis_summary_data to run off data in Zenodo
