+++ pyveg/scripts/analyse_pyveg_summary_data.py
@@ -9,27 +9,33 @@ import os
 import argparse
 import json
 import re
+import tempfile
 import numpy as np
 import pandas as pd
 import seaborn as sns
 import matplotlib.pyplot as plt
 from scipy.stats import pearsonr
 sns.set_style("whitegrid")
-from pyveg.scripts.upload_to_zenodo import upload_results
+
+try:
+    from pyveg.src.zenodo_utils import download_file, list_files, get_deposition_id
+except:
+    print("Unable to import zenodo_utils")
+
 
 def barplot_plots(df, output_dir):
 
     """
     Create barplots of summary data.
-    
-    Parameters 
+
+    Parameters
     -----------
-    df : dataframe 
+    df : dataframe
         Dataframe of summary data.
     output_dir : str
         Path to the directory to save plots to.
     """
-    
+
     plt.figure()
     ax20 = sns.barplot(x='name',y='S2_offset50_mean_max',hue='total_precipitation_mean',data=df)
     ax20.set_xlabel("Mean precipitation over time series")
@@ -47,15 +53,15 @@ def scatter_plots(df, output_dir):
 
     """
     Create scatter plots and correlation plots of summary data.
-    
-    Parameters 
+
+    Parameters
     -----------
-    df : dataframe 
+    df : dataframe
         Dataframe of summary data.
     output_dir : str
         Path to the directory to save plots to.
     """
-    
+
     plt.figure()
     ax = sns.scatterplot(y="S2_offset50_mean_mean", x="total_precipitation_mean", data=df,hue="S2_offset50_mean_pattern_type",palette="Accent_r",edgecolor="k",linewidth=1)
     ax.set_xlabel("Mean precipitation over time series")
@@ -121,7 +127,7 @@ def scatter_plots(df, output_dir):
     ax10.set_ylabel("Offset50 Variance (0.99 rolling window)")
     ax10.set_xlabel("Max Offset50 over time series")
     plt.savefig(os.path.join(output_dir,'offset50max_offset50Variance.png'))
-    
+
     plt.figure()
     ax11 = sns.scatterplot(y="total_precipitation_mean", x="S2_offset50_mean_Variance (0.99 rolling window)", data=df,hue="S2_offset50_mean_pattern_type",palette="Accent_r",edgecolor="k",linewidth=1)
     ax11.set_ylabel("Mean precipitation over time series")
@@ -233,20 +239,35 @@ def boxplot_plots(df, output_dir):
     ax16.set_xlabel("Pattern Type")
     plt.savefig(os.path.join(output_dir, 'Mean_offset50_boxplot.png'))
 
-def process_input_data(input_dir):
+
+def process_input_data(input_location):
 
     """Read all input summary statistics and transform data into
      a more analysis-friendly format
 
     Parameters
     ----------
-    input_dir : str
-        Location of summary statistics output files from analyse_gee_data.py
+    input_location : str
+        Location of summary statistics output files from analyse_gee_data.py.
+        Can be 'zenodo' or 'zenodo_test' to download from the production or sandbox Zenodo depository,
+        or the path to a directory on the local filesystem.
+
+    Returns
+    -------
+    df:  DataFrame containing all the time-series data concatenated together.
     """
+    df_list = []
+    if input_location == 'zenodo' or input_location == 'zenodo_test':
+        is_test = input_location == 'zenodo_test'
+        deposition_id = get_deposition_id(is_test)
+        with tempfile.TemporaryDirectory() as tmpdir:
+            ts_filenames = [download_file(f, deposition_id, tmpdir, is_test) \
+                            for f in list_files(deposition_id, "csv", is_test)]
+            df_list = [pd.read_csv(filename) for filename in ts_filenames]
+    else:
+        ts_filenames = [f for f in os.listdir(input_location) if "_summary_stats" in f]
[…]
