### 2020-10-02 #2483 @usethedata
@crangelsmith Sorry to have been absent, but end of fiscal year stuff was pretty hectic.  I would say you should go ahead and revise the paper.  Let me take a look at the changes in the code from what @arbennett has already review and decide how best to proceed with the review.
