### 2020-10-09 e5aca9f
Update paper.md
Updated authors and acknowledgements. Added in sentences on Zenodo upload and analysis, crystall ball function
