### 2020-10-21 #2483 @usethedata
@crangelsmith Just checking -- what is the timeline for your updates?  I see that you've merged the issue that's related to @arbennett comments, but it looks like the paper hasn't been updated.  No worries, just asking about the timeline.
