### 2020-10-21 #2483 @crangelsmith
Hi @usethedata, sorry for the slow updates.  Currently we are in a transition moment for this project, the RSE funded time  has finished (that is me, @nbarlowATI and @samvanstroud) and we have handed over the finalisation of the project to the researchers in Exeter (@jbuxt, @caboulton and @jabrams23). We will all keep working on getting this paper published but at different capacities.

I think the new version of the paper will be merged into master imminently, but I will let @jbuxt confirm.
