+++ pyveg/paper/paper.bib
@@ -1,3 +1,9 @@
+@misc{Crystal Ball Function,
+    year = {2020},
+    url={https://en.wikipedia.org/wiki/Crystal_Ball_function}
+}
+
+
 @article{Mander:2017,
     author = {Mander, Luke and Dekker, Stefan and Li, Mao and Mio, Washington and Punyasena, Surangi and Lenton, Timothy},
     year = {2017},
