+++ pyveg/paper/paper.bib
@@ -92,3 +92,8 @@
     year = {2020},
     title={Landsat-4, Landsat-5, Landsat-7, Landsat-8 data courtesy of the U.S. Geological Survey}
 }
+
+@misc{Crystal Ball Function,
+    year = {2020},
+    url={https://en.wikipedia.org/wiki/Crystal_Ball_function}
+}
