+++ pyveg/paper/paper.bib
@@ -93,7 +93,7 @@
     title={Landsat-4, Landsat-5, Landsat-7, Landsat-8 data courtesy of the U.S. Geological Survey}
 }
 
-@misc{Crystal Ball Function,
+@misc{CrystalBallFunction,
     year = {2020},
     url={https://en.wikipedia.org/wiki/Crystal_Ball_function}
 }
