+++ pyveg/paper/paper.bib
@@ -94,6 +94,6 @@
 }
 
 @misc{CrystalBallFunction,
-    year = {2020},
+    year = {2016},
     url={https://en.wikipedia.org/wiki/Crystal_Ball_function}
 }
