### 2020-10-22 #2483 @jbuxt
Hello @usethedata apologies for the delay in merging the paper into the master. It should all be up to date now, please do let me know if there are any comments/questions.
