### 2020-10-24 #2483 @arbennett
I've re-looked at things and several of my initial issues have been solved, but it looks like work hasn't quite completed on all of them. The updated documentation looks really great and is very in depth - thank you for that! It seems the tutorials and examples still haven't quite been updated to run through smoothly. From what I can see this is still a known issue, as there are open issues on these topic, so I'm leaving these unchecked in my initial review.

On the paper itself, I think it explains the motivation and overall structure of `pyveg` quite well, but as someone who does not work on ecosystem resilience I'm not quite ready to check the 'State of the Field' box. I think it would be really helpful to detail any previous tools or initiatives for monitoring ecosystem resiliency. If there really haven't been any, I think that's also worth stating.

I also have one minor point on the referencing that I think would be nice to update, the part on the Crystal Ball distribution. Instead of linking to the Wikipedia page here, is there a reference where this distribution has been used for this purpose? That would tie in better, and give more of a springboard for interested parties.
