### 2020-10-24 #2483 @usethedata
@arbennett I do work in and around ecosystem resiliency and work with others who do.  I'll take a particular look at that part.
