### 2020-10-25 #2483 @usethedata
Minor issue with paper bibliography case.  Fix proposed in https://github.com/alan-turing-institute/monitoring-ecosystem-resilience/pull/481
