### 2020-10-25 #2483 @usethedata
An observation for the team @crangelsmith @samvanstroud is that the Azure messages feel annoying and are inconsistent with the structure of the other output.  For example, getting this output from the test run:

```
    azure_config.py not found - this is needed for using Azure storage or batch.
    Copy pyveg/azure_config_template.py to pyveg/azure_config.py then input your
    own values for Azure Storage account name and Access key, then redo `pip install .`

    azure_config.py not found - this is needed for using Azure storage or batch.
    Copy pyveg/azure_config_template.py to pyveg/azure_config.py then input your
    own values for Azure Storage account name and Access key, then redo `pip install .`

2020-10-25 14:36:31,897 [INFO] Sentinel2: setting collection_name to COPERNICUS/S2
2020-10-25 14:36:31,897 [INFO] Sentinel2: setting data_type to vegetation
```

getting the azure_config.py message twice, and not as a (for example) `[INFO]` message feels off.  For that matter, why not have the default install be something that includes a `azure_config.py` file that says "I don't have Azure access and don't bug me about it" and then provide the instructions for people to change that.  Or, at least, why pester the user about that file being missing?  It's a bit of a small thing, but in an Open Source context, nagging users about not using a commercial, paid-for service seems a bit of a blemish on some really great work.  Supporting Azure is great, and no worries about that.  It's just the default behavior that feels a bit off.
