### 2020-10-25 #2483 @usethedata
Regarding the state of the field question @arbennett raises:  I've read the section of the paper carefully and consulted with a couple of colleagues on the question.  The short answer is that I'm comfortable with this as presented.

The longer answer is that the ecosystem resilience is an application area as far as this paper is concerned.  The paper does not have an exhaustive list of references on the topic, but that's also not necessary in my view (as a still relatively new JOSS topic editor).  They point to a couple of papers which my colleagues felt were reasonably representative of the subject and provide sufficient summary of the are for a reader interested in that particular application area.  This is a paper about the software, and would support, for example a paper in the *Journal of Remote Sensing* or other venue to discuss more of the science.

I also considered the question from the state of the field perspective for the algorithms presented, and I'm also comfortable that this was done at a level consistent with other JOSS papers that I went out to look at.  So, I'm comfortable checking that box from my perspective.  @arbennett  -- does that address your question?
