### 2020-10-26 #2483 @jbuxt
@arbennett and @usethedata Thank you both for reviewing this over the weekend and for providing constructive and helpful comments.

I believe that @crangelsmith has approved the pull request to update the biblography and I will improve the Crystal ball reference. We will discuss as a group how best to address the rest of the comments in order to make some improvements.
