### 2020-10-26 ed5e5bd
Update paper.md
Added Chris' initial
