### 2020-10-26 #2483 @arbennett
@usethedata  - thanks for checking that. I'm happy to check the box then given your perspective. I'm also pretty new to reviewing for JOSS so I'm not always certain on the exact criteria. This is helpful though, thanks!
