### 2020-10-27 #2483 @usethedata
@jbuxt -- as a clarification to my earlier comments (about the Azure message).  If there's a simple fix for that issue, then great, and I'd love to see it addressed before publication.  However, I do not consider it a blocker from my perspective as a reviewer.  It is a suggestion for the team to consider.
