### 2020-10-27 #2483 @nbarlowATI
@usethedata I think it is a good point about the Azure message being a bit "nag"y, when as you say, it is entirely optional whether the user would want to use Azure.   We will silence this (probably relegate it to an ERROR if a function that actually does use Azure is called) in the next update, in the next couple of days.
