### 2020-10-28 #2483 @arbennett
I think the paper itself is in good condition but still wasn't able to successfully run through the tutorials via binder.  It seems issue I submitted is still open, so I think that's still being worked out.
