### 2020-11-06 #2483 @usethedata
@crangelsmith @nbarlowATI  I believe @arbennett is referring to https://github.com/alan-turing-institute/monitoring-ecosystem-resilience/issues/367, which is still showing open, but I see a note that it was moved to Done on September 18th. But, given that @arbennett commented on seeing this issue recently, it doesn't feel closed.  Can you help me understand what's going on?  Thanks much.
