### 2020-11-09 2ec5915
add warning when lagged correlation plots not present, rather than raising error
