+++ pyveg/src/plotting.py
@@ -99,15 +99,17 @@ def plot_time_series(df, output_dir):
             ax2.plot(precip_xs, precip_ys, linewidth=2, color=color, alpha=0.75)
 
             # add veg-precip correlation
-            max_corr_smooth, max_corr = get_max_lagged_cor(os.path.dirname(output_dir), veg_prefix)
-            textstr = f'$r_{{t-{max_corr[1]}}}={max_corr[0]:.2f}$'
+            try:
+                max_corr_smooth, max_corr = get_max_lagged_cor(os.path.dirname(output_dir), veg_prefix)
+                textstr = f'$r_{{t-{max_corr[1]}}}={max_corr[0]:.2f}$'
 
-            # old correlation just calculates the 0-lag correlation
-            #raw_corr = veg_means.corr(precip_ys)
-            #smoothed_corr = veg_means_smooth.corr(precip_ys)
-            #textstr = f'$r={smoothed_corr:.2f}$ (${raw_corr:.2f}$ unsmoothed)'
-            ax2.text(0.13, 0.95, textstr, transform=ax2.transAxes, fontsize=14, verticalalignment='top')
-
+                # old correlation just calculates the 0-lag correlation
+                #raw_corr = veg_means.corr(precip_ys)
+                #smoothed_corr = veg_means_smooth.corr(precip_ys)
+                #textstr = f'$r={smoothed_corr:.2f}$ (${raw_corr:.2f}$ unsmoothed)'
+                ax2.text(0.13, 0.95, textstr, transform=ax2.transAxes, fontsize=14, verticalalignment='top')
+            except(FileNotFoundError):
+                print("Warning - lagged correlations file not found - call plot_autocorrelation_function() first to print correlations on plot")
         # plot second vegetation time series if availible
         if veg_prefix_b:
 
@@ -137,11 +139,11 @@ def plot_time_series(df, output_dir):
             ax.plot(veg_xs_b, veg_means_b, label='Unsmoothed', linewidth=1, color='indigo', linestyle='dashed', alpha=0.2)
 
             # plot smoothed vegetation means and std
-            ax3.plot(veg_xs_b, veg_means_smooth_b, marker='o', markersize=7, 
-                    markeredgecolor=(0.8172, 0.7627, 0.9172), markeredgewidth=2, 
+            ax3.plot(veg_xs_b, veg_means_smooth_b, marker='o', markersize=7,
+                    markeredgecolor=(0.8172, 0.7627, 0.9172), markeredgewidth=2,
                     label='Smoothed', linewidth=2, color=color)
 
-            ax3.fill_between(veg_xs_b, veg_means_smooth_b - veg_stds_smooth_b, veg_means_smooth_b + veg_stds_smooth_b, 
+            ax3.fill_between(veg_xs_b, veg_means_smooth_b - veg_stds_smooth_b, veg_means_smooth_b + veg_stds_smooth_b,
                             facecolor='tab:purple', alpha=0.1, label='Std Dev')
 
             # add veg-veg correlation
@@ -216,12 +218,12 @@ def plot_ndvi_time_series(df, output_dir):
         # plot ndvi
         #ax.plot(veg_xs, ndvi_means, label='Unsmoothed', linewidth=1, color='dimgray', linestyle='dotted')
 
-        ax.plot(veg_xs, veg_means, marker='o', markersize=7, 
+        ax.plot(veg_xs, veg_means, marker='o', markersize=7,
                 markeredgecolor=(0.9172, 0.9627, 0.9172), markeredgewidth=2,
                 label='Smoothed', linewidth=2, color='green')
 
         if any([col_name.replace('mean', 'std') == c for c in df.columns]):
-            ax.fill_between(veg_xs, veg_means - veg_std, veg_means + veg_std, 
[…]
