### 2020-11-09 e65204c
remove warning about Zenodo config
