### 2020-11-09 1aec618
add option to not plot smoothed time-series
