### 2020-11-09 #2483 @nbarlowATI
Hi @usethedata I believe that this particular notebook / cell that @arbennett referred to in  alan-turing-institute/monitoring-ecosystem-resilience#367 was fixed back then (that notebook has been completely rewritten, and the problem of download sizes being exceeded is fixed).   I just tested the latest version on Binder and it does now run through (although the cell that runs the pipeline is extremely slow - I guess Binder (understandably) don't provide super-fast VMs!)

There were some problems in the other tutorials notebooks though, which I believe may have been what @arbennet was referring to in the recent comment.   These should hopefully also be fixed now in the latest version pushed to `master` today.
