### 2020-11-10 3244493
add argument to upload_to_zenodo script to allow csv upload
