### 2020-11-19 #2483 @arbennett
I tried out the updated tutorial, and things ran through, though it took a bit of time. Still, that was my only remaining stumbling block, so I would be happy to accept this submission!
