### 2020-11-19 #2483 @usethedata
@samvanstroud (and team):
1) Looking at the repository, I'm seeing version 1.0 as the current release, with a date of March 2020.  I don't know if you want to tag at 1.0.1 or something similar to note the changes made since this review started.  Please confirm what you want for the version number for the paper and either let me know you're going with 1.0 or set up the new release/tag and let me know that.
2) When you have that, please create a deposit at Zenodo and post back here with the DOI that they give you.

Regards,
Bruce
