### 2020-11-19 b536d35
bump version number to 1.1.0
