### 2020-11-19 #2483 @crangelsmith
Hi @usethedata,

> 1. Looking at the repository, I'm seeing version 1.0 as the current release, with a date of March 2020.  I don't know if you want to tag at 1.0.1 or something similar to note the changes made since this review started.  Please confirm what you want for the version number for the paper and either let me know you're going with 1.0 or set up the new release/tag and let me know that.

Yes, we decided to move to version v1.1.0, this is now tagged and updated on the [repo](https://github.com/alan-turing-institute/monitoring-ecosystem-resilience/releases/tag/v1.1.0).

>2. When you have that, please create a deposit at Zenodo and post back here with the DOI that they give you.

Done, this is the DOI they provide: [10.5281/zenodo.4281273](https://doi.org/10.5281/zenodo.4281273)

Thank you and @arbennett for all your work reviewing this paper!!!
